### 2019-10-09 #1792 @whedon
[PRE REVIEW]: archetypal: A Python package for collecting, simulating, converting and analyzing building archetypes
**Submitting author:** @samuelduchesne (<a href="http://orcid.org/0000-0001-5790-878X">Samuel Letellier-Duchesne</a>)
**Repository:** <a href="https://github.com/samuelduchesne/archetypal" target ="_blank">https://github.com/samuelduchesne/archetypal</a>
**Version:** v1.3
