### 2019-10-09 #1792 @labarba
👋 @samuelduchesne — I am the Associate Editor-in-Chief on rotation this week, and will continue working to identify an editor to handle your submission.  But first, I need to ask you to add to the top of your paper:

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

(From our [Documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).)

Currently, the paper is using a lot of jargon. As it is, I (a non-specialist) will find it hard for to identify the right editor. Thanks!
