### 2019-10-09 77701e1
Adds a clearer message in the summary and removes unnecessary acronyms
Also fixes heading names
