+++ paper/paper.md
@@ -23,48 +23,54 @@ bibliography: paper.bib
 
 # Summary
 
-`archetypal` is a Python package that helps handle collections of building archetypes. It
-offers 3 majors functionalities:
+
+
+The field of Urban Building Energy Modeling, which assesses the energy performance of
+buildings in cities relies on advanced physical models known as Building Energy Models
+(BEM) or simply, building archetypes. *archetypal* is a Python package that helps handle
+collections of such archetypes developed on the EnergyPlus model architecture. It offers 3
+majors capabilities:
 
 1. Run, modify and analyze collections of EnergyPlus models in a persistent environment;
-2. Convert [EnergyPlus](https://energyplus.net) IDF models to [UMI Template Files](http://web.mit.edu/sustainabledesignlab/projects/umi/index.html);
-3. Convert [EnergyPlus](https://energyplus.net) IDF models to TRNSYS [TrnBuild](http://www.trnsys.com/features/suite-of-tools.php) BUI Models.
+2. Convert [EnergyPlus](https://energyplus.net) models to [UMI Template Files](http://web.mit.edu/sustainabledesignlab/projects/umi/index.html);
+3. Convert [EnergyPlus](https://energyplus.net) models to TRNSYS [TrnBuild](http://www.trnsys.com/features/suite-of-tools.php) BUI Models.
  
 ## EnergyPlus Simulation Environment
 
-`archetypal` leverages the Python Eppy [@Philip2004] and GeomEppy [@Bull2016] packages to
+*archetypal* leverages the Python Eppy [@Philip2004] and GeomEppy [@Bull2016] packages to
 handle parsing and modifications of EnergyPlus files. Additional functionalities where
 developed such as a caching system and a file upgrade system as well as other class
-methods and properties that are specific to building archetype analysis. `archetypal` lets
+methods and properties that are specific to building archetype analysis. *archetypal* lets
 users query EnergyPlus results to return specific time series in a DataFrame format. For
 convenience, useful time series such as the space heating, space cooling and domestic hot
 water profiles are accessible by default. Users can also specify other output names and
-`archetypal` will append the IDF file and rerun the simulation.
+*archetypal* will append the IDF file and rerun the simulation.
 
-Furthermore, `archetypal` features a caching method that handles simulation results. This
+Furthermore, *archetypal* features a caching method that handles simulation results. This
 is particularly useful for reproducible workflows such as the Jupyter Notebook programing
-environment. Reopening a closed notebook and running a cell containing the `run` command
-will use the cached simulation results instead of executing EnergyPlus again. This offers
-a drastic workflow speed gain especially when larger IDF files can take several minutes to
-complete.
+environment. Reopening a closed notebook and running a cell containing the `run_eplus`
+command will use the cached simulation results instead of executing EnergyPlus again. This
+offers a drastic workflow speed gain especially when larger IDF files can take several
+minutes to complete.
 
-## EnergyPlus to UMI Template
+## EnergyPlus to UMI Template File Conversion
 
-`archetypal` aims at providing a way of creating UMI Template Files from EnergyPlus
-models. The algorithm approximates the non-geometric parameters of a multi-zone EnergyPlus
-model by dissecting and combining core zones and perimeter zones. The procedure is an
-attempt to streamline the creation of Urban Building Energy Models (UBEM) [@Reinhart2016] based on the
-"Shoeboxer" method [@Dogan2017] by accelerating the creation of building archetype
+*archetypal* aims at providing a way of creating UMI Template Files from EnergyPlus models.
+The algorithm approximates the non-geometric parameters of a multi-zone EnergyPlus model
+by dissecting and combining core zones and perimeter zones. The procedure is an attempt to
[…]
