### 2019-10-09 475b746
Moves figure bellow text
