### 2019-10-09 #1792 @labarba
@brynpickering 👋  — Would you be interested in contributing a review for The Journal of Open Source Software (JOSS) on this submission:
**archetypal: A Python package for collecting, simulating, converting and analyzing building archetypes**
https://github.com/samuelduchesne/archetypal
