### 2019-10-09 #1792 @samuelduchesne
🤓 Hi @labarba. Thanks for handling this review process. I looked at the list of reviewers and filtered it for python users and the keyword "energy" in their topic area (nothing for *building*). Only 9 hits! Most of them specialists in physics. May I suggest @katyhuff, @wholmgren or @EricaEgg? @brynpickering was also in the list.
