### 2019-10-10 #1792 @labarba
Thank you, @brynpickering !

The area of application of this software is unusual for JOSS, so I think finding reviewers could be a bit of a challenge. Would you be able to recommend another person to review this?
