### 2019-10-10 #1792 @brynpickering
Hmm, that's a tricky one - most people I know in the field are more accustomed to Matlab, so the Python element might prove to be a barrier. Nonetheless, @rmward61 might be interested (/ know others).
