### 2019-10-13 #1792 @labarba
@katyhuff — Could you handle this submission as editor? We already have one reviewer and need a second.
archetypal: A Python package for collecting, simulating, converting and analyzing building archetypes
https://github.com/samuelduchesne/archetypal
(related to assessing energy performance of buildings in cities)
