### 2019-10-22 #1792 @Kevin-Mattheus-Moerman
@kyleniemeyer, @ThibaultMarzullo is a colleague of mine at NUIG and works in this domain.

@ThibaultMarzullo perhaps you can also update your GitHub profile to list your academic affiliation and perhaps even an ORCID profile or similar. Thank for your help here!
