### 2019-10-29 e57332f
New: Add argument to be able to search a text file from a given line
