### 2019-10-29 e3cf623
New test to modify b18 file
