### 2019-10-29 120a548
New: Adds infiltration to b18 file
