### 2019-10-29 09b11bb
Fix: Selects the used schedule instead of returning an empty list
