### 2019-10-30 92ab1b1
Only writes yearly schedules
