### 2019-10-30 b2949b8
New: Keeps in memory the unwritten schedules
