### 2019-10-30 0fd534a
New: Adds ZONELIST object
