### 2019-10-30 #1833 @ThibaultMarzullo
Hi @samuelduchesne, I'm going through the paper and the software now. It's an interesting idea and work.

**Paper:**
It would benefit from the addition of some context as well as a couple of clarifications, see [this issue](https://github.com/samuelduchesne/archetypal/issues/41) for more details.

**Documentation:**
Generally speaking it seems incomplete, the user is left to finding out on their own how to use _archetypal_ for any task other than converting models. See [this issue](https://github.com/samuelduchesne/archetypal/issues/41) for more details.

I'll get back to you as soon as I run the tests, the only issue is that I won't be able to test the TRNSYS model since I don't own a license.
