### 2019-10-30 samuelduchesne/archetypal#41 @louisleroy5
Hello Thibault,

Thank you very much for the review !
I find the comments really constructive. I'll try to put some time this week to improve the documentation. Then I'll work on the paper.

Best regards,
Louis
