### 2019-10-30 e17c3d5
Adds weather file input to be able to run EP
