+++ tests/test_trnsys.py
@@ -77,22 +77,26 @@ class TestsConvert:
         }
         idf = load_idf(file)
 
-        yield idf, file, window_filepath, trnsidf_exe, template_d18, kwargs_dict
+        weather_file = os.path.join(
+            "tests", "input_data", "CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw"
+        )
+
+        yield idf, file, weather_file, window_filepath, trnsidf_exe, template_d18, kwargs_dict
 
         del idf
 
     def test_load_idf_file_and_clean_names(self, config, converttest):
-        idf, idf_file, window_lib, trnsidf_exe, template, _ = converttest
+        idf, idf_file, weather_file, window_lib, trnsidf_exe, template, _ = converttest
         log_clear_names = False
         idf_2 = _load_idf_file_and_clean_names(idf_file, log_clear_names)
 
     def test_get_save_write_schedules(self, config, converttest):
         output_folder = None
-        idf, idf_file, window_lib, trnsidf_exe, template, _ = converttest
+        idf, idf_file, weather_file, window_lib, trnsidf_exe, template, _ = converttest
         lines = io.TextIOWrapper(io.BytesIO(settings.template_BUI)).readlines()
         try:
-            idf_file, window_lib, output_folder, trnsidf_exe, template = _assert_files(
-                idf_file, window_lib, output_folder, trnsidf_exe, template
+            idf_file, weather_file, window_lib, output_folder, trnsidf_exe, template = _assert_files(
+                idf_file, weather_file, window_lib, output_folder, trnsidf_exe, template
             )
         except:
             output_folder = os.path.relpath(settings.data_folder)
@@ -102,7 +106,7 @@ class TestsConvert:
         schedules_not_written = _write_schedules(lines, schedule_names, schedules)
 
     def test_write_version_and_building(self, config, converttest):
-        idf, idf_file, window_lib, trnsidf_exe, template, _ = converttest
+        idf, idf_file, weather_file, window_lib, trnsidf_exe, template, _ = converttest
         buildingSurfs, buildings, constructions, equipments, fenestrationSurfs, globGeomRules, lights, locations, materialAirGap, materialNoMass, materials, peoples, versions, zones, zonelists = _get_idf_objects(
             idf
         )
@@ -111,7 +115,9 @@ class TestsConvert:
         _write_building(buildings, lines)
 
     def test_write_idf_objects(self, config, converttest):
-        idf, idf_file, window_lib, trnsidf_exe, template, kwargs = converttest
+        idf, idf_file, weather_file, window_lib, trnsidf_exe, template, kwargs = (
+            converttest
+        )
 
         # Read IDF_T3D template and write lines in variable
         lines = io.TextIOWrapper(io.BytesIO(settings.template_BUI)).readlines()
@@ -199,7 +205,7 @@ class TestsConvert:
         _write_winPool(lines, window)
 
     def test_write_material(self, config, converttest):
-        idf, idf_file, window_lib, trnsidf_exe, template, _ = converttest
+        idf, idf_file, weather_file, window_lib, trnsidf_exe, template, _ = converttest
 
         # Read IDF_T3D template and write lines in variable
         lines = io.TextIOWrapper(io.BytesIO(settings.template_BUI)).readlines()
@@ -213,7 +219,27 @@ class TestsConvert:
         _write_materials(lines, materialAirGap, materialNoMass, materials)
 
     def test_write_gains(self, config, converttest):
-        idf, idf_file, window_lib, trnsidf_exe, template, _ = converttest
[…]
