+++ tests/test_trnsys.py
@@ -147,6 +147,7 @@ class TestsConvert:
             materials,
             peoples,
             zones,
+            zonelists,
             ordered,
         )
 
