### 2019-10-30 2c20be0
Adds run_eplus + store equivalence between old and new names
