### 2019-10-31 08d7fbd
New: Gets internal gains from 'htm' simulation results
