+++ archetypal/trnsys.py
@@ -214,7 +214,7 @@ def convert_idf_to_trnbuild(
     # endregion
 
     # region Write GAINS (People, Lights, Equipment) from IDF to lines (T3D)
-    _write_gains(equipments, idf, lights, lines, peoples)
+    _write_gains(equipments, idf, lights, lines, peoples, res, old_new_names)
     # endregion
 
     # region Write SCHEDULES from IDF to lines (T3D)
@@ -1821,7 +1821,7 @@ def _write_schedules(lines, schedule_names, schedules):
     return schedules_not_written
 
 
-def _write_gains(equipments, idf, lights, lines, peoples):
+def _write_gains(equipments, idf, lights, lines, peoples, res, old_new_names):
     """Write gains in lines
 
     Args:
@@ -1838,14 +1838,14 @@ def _write_gains(equipments, idf, lights, lines, peoples):
     # Get line number where to write
     gainNum = checkStr(lines, "G a i n s")
     # Writing PEOPLE gains infos to lines
-    _write_people_gain(gainNum, idf, lines, peoples)
+    _write_people_gain(gainNum, lines, peoples, res, old_new_names)
     # Writing LIGHT gains infos to lines
-    _write_light_gain(gainNum, lights, lines)
+    _write_light_gain(gainNum, lights, lines, res, old_new_names)
     # Writing EQUIPMENT gains infos to lines
-    _write_equipment_gain(equipments, gainNum, lines)
+    _write_equipment_gain(equipments, gainNum, lines, res, old_new_names)
 
 
-def _write_equipment_gain(equipments, gainNum, lines):
+def _write_equipment_gain(equipments, gainNum, lines, res, old_new_names):
     """Write equipment gains in lines
 
     Args:
@@ -1855,60 +1855,30 @@ def _write_equipment_gain(equipments, gainNum, lines):
         lines (list): Text to create the T3D file (IDF file to import in
             TRNBuild). To be appended (insert) here
     """
-    for i in range(0, len(equipments)):
-        # Determine if gain is absolute or relative and write it into lines
-        if equipments[i].Design_Level_Calculation_Method == "EquipmentLevel":
-            areaMethod = "ABSOLUTE"
-            try:
-                power = round(float(equipments[i].Design_Level), 4)
-            except Exception:
-                log(
-                    "Could not find the Light Power Density in the IDF file", lg.WARNING
-                )
-                continue
-        elif equipments[i].Design_Level_Calculation_Method == "Watts/Area":
-            areaMethod = "AREA_RELATED"
-            try:
-                power = round(float(equipments[i].Watts_per_Zone_Floor_Area), 4)
-            except Exception:
-                log(
-                    "Could not find the Light Power Density in the IDF file", lg.WARNING
-                )
-                continue
-        else:
-            log(
-                "Could not find the Equipment Power Density, cause depend on "
-                "the number of peoples (Watts/Person)",
-                lg.WARNING,
+    for equipment in equipments:
+        gain = res["ElectricEquipment Internal Gains Nominal"][
+            res["ElectricEquipment Internal Gains Nominal"]["Name"].str.contains(
+                old_new_names[equipment.Name.upper()][0]
             )
-            continue
-
-        # Find the radiant fractions
-        radFract = equipments[i].Fraction_Radiant
-        if len(str(radFract)) == 0:
-            # Find the radiant fractions
-            try:
-                radFract = float(1 - equipments[i].Sensible_Heat_Fraction)
-            except Exception:
-                radFract = 0.42
-        else:
-            radFract = float(radFract)
-
-        # Write gain from equipment in lines
[…]
