### 2019-10-31 ac25abc
Equivalence new/old names has new_names.upper() as keys and old_names.upper() as values
