### 2019-10-31 ff34db1
Fixes when checking in lines from a given line number
