### 2019-10-31 3cd8cfc
Fix: delete one unnecessary argument
