### 2019-10-31 f12f72c
New: writes internal gains in b18
