### 2019-10-31 e616e40
New: writes infiltration in zone's sections
