### 2019-10-31 3b0c126
Fixes: makes sure hours_list is not empty
