### 2019-10-31 0db2835
Fix: delete one unnecessary argument
