### 2019-10-31 43db14b
Adds idf name to old/new names equivalence csv file
