### 2019-11-01 fc3fd6e
Makes sure output folder exists
