### 2019-11-01 46fab8f
Cleans names on a copy of the idf, like that run_eplus on the original idf
