+++ archetypal/trnsys.py
@@ -27,11 +27,12 @@ from archetypal import (
     load_idf_object_from_cache,
     hash_file,
     run_eplus,
+    copy_file,
 )
 from geomeppy.geom.polygons import Polygon3D
 from path import Path
 from tqdm import tqdm
-
+from copy import copy
 
 def convert_idf_to_trnbuild(
     idf_file,
@@ -119,7 +120,20 @@ def convert_idf_to_trnbuild(
     )
 
     # Check if cache exists
-    idf = _load_idf_file_and_clean_names(idf_file, log_clear_names)
+    # idf = _load_idf_file_and_clean_names(idf_file, log_clear_names)
+    idf = load_idf(idf_file)
+
+    # Clean names of idf objects (e.g. 'MATERIAL')
+    idf_2 = copy(idf)
+    log("Cleaning names of the IDF objects...", lg.INFO)
+    start_time = time.time()
+    clear_name_idf_objects(idf_2, log_clear_names)
+    log(
+        "Cleaned IDF object names in {:,.2f} seconds".format(
+            time.time() - start_time
+        ),
+        lg.INFO,
+    )
 
     # Get old:new names equivalence
     old_new_names = pd.read_csv(
@@ -134,7 +148,7 @@ def convert_idf_to_trnbuild(
 
     # Get objects from IDF file
     buildingSurfs, buildings, constructions, equipments, fenestrationSurfs, globGeomRules, lights, locations, materialAirGap, materialNoMass, materials, peoples, versions, zones, zonelists = _get_idf_objects(
-        idf
+        idf_2
     )
 
     # Get all construction EXCEPT fenestration ones
@@ -162,13 +176,13 @@ def convert_idf_to_trnbuild(
     )
 
     # region Get schedules from IDF
-    schedule_names, schedules = _get_schedules(idf)
+    schedule_names, schedules = _get_schedules(idf_2)
 
     _yearlySched_to_csv(idf_file, output_folder, schedule_names, schedules)
     # endregion
 
     # Gets and removes from IDF materials with resistance lower than 0.0007
-    mat_name = _remove_low_conductivity(constructions, idf, materials)
+    mat_name = _remove_low_conductivity(constructions, idf_2, materials)
 
     # Write data from IDF file to T3D file
     start_time = time.time()
@@ -186,11 +200,11 @@ def convert_idf_to_trnbuild(
     coordSys = _is_coordSys_world(coordSys, zones)
 
     # Change coordinates from relative to absolute for building surfaces
-    _change_relative_coords(buildingSurfs, coordSys, idf)
+    _change_relative_coords(buildingSurfs, coordSys, idf_2)
 
     # Adds or changes adjacent surface if needed
-    _add_change_adj_surf(buildingSurfs, idf)
-    buildingSurfs = idf.idfobjects["BUILDINGSURFACE:DETAILED"]
+    _add_change_adj_surf(buildingSurfs, idf_2)
+    buildingSurfs = idf_2.idfobjects["BUILDINGSURFACE:DETAILED"]
 
     # region Write VARIABLEDICTONARY (Zone, BuildingSurf, FenestrationSurf)
     # from IDF to lines (T3D)
@@ -201,16 +215,16 @@ def convert_idf_to_trnbuild(
 
     # Writing zones in lines
     win_slope_dict = _write_zone_buildingSurf_fenestrationSurf(
-        buildingSurfs, coordSys, fenestrationSurfs, idf, lines, n_ground, zones
+        buildingSurfs, coordSys, fenestrationSurfs, idf_2, lines, n_ground, zones
     )
     # endregion
 
     # region Write CONSTRUCTION from IDF to lines (T3D)
[…]
