### 2019-11-01 5109909
Takes off ApatmentMidrise from test, because problem when running
