### 2019-11-05 samuelduchesne/archetypal#42 @brynpickering
[Request] Document conda installation workflow
Installing on MacOS fails to complete due to a pip wheel not being found for 'regex' (one of the dependencies of a dependency). This isn't an issue with archetypal, per se, but was solved by a conda installation workflow (i.e. creating a conda environment, installing all dependencies from conda / conda forge, then installing archetypal with pip). Would it be possible to:

a) have a working conda environment requirements file in this repository, to minimise direct pip dependency, and

b) update the docs to include information on installing archetypal using a conda workflow, instead of directly with pip?

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/1833)

Edit: I see there is already a requirements file, which is used in the Docker instance, but it would need updating to work with MacOS too...
