### 2019-11-05 samuelduchesne/archetypal#43 @brynpickering
[Request] Better management of EnergyPlus dependency
The docs specify E+ v.8.9.0, when v.9.2.0 is already available (I can safely say that `archetypal` doesn't work with v.9.2.0!). Version issues crop up relatively late on in the use of `archetypal`, and are not particularly explicit in providing advice on how to solve them in the error messages. Could you check for E+ version on python interactive import of `archetypal`, then raise errors then about any user E+ version issues, instead of waiting for one of the `archetypal` functions to fail later on? It might also help to make the version dependency even more explicit in e.g. the package README.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/1833)
