### 2019-11-05 samuelduchesne/archetypal#44 @brynpickering
[Docs] Incorrect example code
This example code from `Convert IDF to UMI` doesn't work:
```
from archetypal import BuildingTemplate
template_obj = BuildingTemplate.from_idf(
    x.idf, sql=x.idf.sql, DataSource=x.idf.name
)
```
^ Variable `x` is not yet defined elsewhere in the page

```
from archetypal import UmiTemplate
template_json = UmiTemplate(
     name="my_umi_template",
     BuildingTemplates=template_obj
).to_json()
```
^ `template_obj` is expected to be iterable. In the docs example, only one file is loaded, so `template_obj` (assuming `x.` is removed) is only one, uniterable object.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/1833)
