### 2019-11-05 samuelduchesne/archetypal#45 @brynpickering
Tests failing on MacOS
8 tests fail and 22 error when running tests locally on MacOS. The 22 errors seem to all be trnsys related, but could be handled more gracefully by skipping a test if the operating system is not Windows (see [here](http://doc.pytest.org/en/latest/skipping.html)). Full test log is given [here](https://gist.github.com/brynpickering/569274c474813f72e6730e35f3d5ad99) - enjoy!

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/1833)
