### 2019-11-05 samuelduchesne/archetypal#46 @brynpickering
archetypal convert error
On running ``archetypal convert "SimpleOfficeBuilding.idf" "output"`` I get the following error:

```
Traceback (most recent call last):
  File "my_conda_env/bin/archetypal", line 8, in <module>
    sys.exit(cli())
  File "my_conda_env/lib/python3.7/site-packages/click/core.py", line 764, in __call__
    return self.main(*args, **kwargs)
  File "my_conda_env/lib/python3.7/site-packages/click/core.py", line 717, in main
    rv = self.invoke(ctx)
  File "my_conda_env/lib/python3.7/site-packages/click/core.py", line 1137, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "my_conda_env/lib/python3.7/site-packages/click/core.py", line 956, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "my_conda_env/lib/python3.7/site-packages/click/core.py", line 555, in invoke
    return callback(*args, **kwargs)
  File "my_conda_env/lib/python3.7/site-packages/archetypal/cli.py", line 303, in convert
    if paths:
UnboundLocalError: local variable 'paths' referenced before assignment
```

It looks like in `cli.py` there is a `with` statement which produces a variable `paths` which is then expected to exist later on (`if paths:`). `paths` might need to be set as e.g. None before the `with` statement?

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/1833)
