### 2019-11-05 samuelduchesne/archetypal#47 @brynpickering
JOSS paper review
Just two points on your JOSS paper (otherwise, I think it is sufficiently clear):

1. Line 62: striped -> stripped

2. Lines 79-81: Could you add links to the references to DOE benchmark models, to allow interested users to easily access those models?

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/1833)
