### 2019-11-05 #1833 @brynpickering
Hey @samuelduchesne, just had a chance to go through this. A tool to help collate building archetypes and work on collections of models seems incredibly useful, but I've come across a few issues that would need addressing to make sure it is a sufficiently robust piece of software for general use. It looks like neither myself or @ThibaultMarzullo will be able to test the TRNSYS functionality, so some demonstration from your end might be necessary to convince us on that...?

MacOS compatibility
---------------------
see https://github.com/samuelduchesne/archetypal/issues/42 and https://github.com/samuelduchesne/archetypal/issues/45

I'm not sure this has actually been tested on MacOS. The installation instructions do cover it, but I have hit a few roadblocks along the way, and some of the inbuilt tests fail. I would consider a conda forge version of archetype, where you can do OS-specific testing, for a more robust package.

EnergyPlus compatibility
-------------------------
see https://github.com/samuelduchesne/archetypal/issues/43

Given that E+ is such an integral part of your software, it would be good to have a better handle on allowed E+ versions and dealing with a. IDF files without explicit versions and b. use of v.9.2.0 (current latest E+ version, currently fails).

Documentation
---------------
see https://github.com/samuelduchesne/archetypal/issues/44

This is currently seriously lacking, with errors in the little example code that there is. There seems to be *a lot* of functionality in the software, but it is all lost due to lack of tutorial documentation, and limited API method documentation. Using e.g. in-built E+ example models / DOE benchmark models to show off loading in multiple files, comparing DataFrames, comparing solve times, etc. is really necessary to help users understand the software and its benefits. As is, I don't feel able to really play around with the software as I don't know what most of the possibilities are without delving into each function in the core code.

Tests and graceful failure
--------------------------
see https://github.com/samuelduchesne/archetypal/issues/46, https://github.com/samuelduchesne/archetypal/issues/43 and https://github.com/samuelduchesne/archetypal/issues/45

I think tests are currently lacking, since I've seemingly broken the software already on multiple occasions (some of which is MacOS-specific). In some instances, I would expect to see a much clearer error message (e.g. E+ version issues), which requires catching and processing some of the more likely strange ways a user could abuse the software. But in other cases, the package fails when following the documentation...

Paper
------
see https://github.com/samuelduchesne/archetypal/issues/47
I'm happy with the paper, I get a good grasp of what I could do with archetypal, and you cite sources for the methods which underpin what you do. It might be nice to see a table of performance metrics to compare running a new E+ model, cached E+ model, and 'shoeboxed' E+ model, to show the relevant benefits of key archetypal functionality.
