### 2019-11-05 44d80fb
Makes sure zone coords are not empty strings
