### 2019-11-06 7f66e19
New test for simple IDF file
