### 2019-11-06 cc30394
Move part of code that bug if before "versionid == to_version"
