### 2019-11-06 a7ddb94
Fixes internal gain category
