### 2019-11-06 4c40a2c
Corrects internal gain names
