### 2019-11-06 7932108
Makes sure "idf_file" is a Path()
