### 2019-11-06 2749f02
Writes whole schedule in lines (even too long ones) - TRNBuild crashes when to much long schedules
