### 2019-11-06 d319b6d
Needs to run ExpandObject for ZoneHVAC:Template objects
