### 2019-11-06 c5c71ee
Uses ExpandObjects for ZoneHVAC:Template objects
