### 2019-11-07 a9f7d59
Writes conditioning (heating and cooling) in T3D and B18 file
