### 2019-11-07 a6e7833
Makes sure ElecPowerFrac=1 and Humidity=0 for light and equip
