### 2019-11-07 c52d104
Makes sure internal gain refers to AREA
