### 2019-11-08 e5a538a
adds schedule name, if available, as label in plot
