### 2019-11-08 d2a6844
Adds quantity parameter to schedule class combine method
