### 2019-11-08 0ad6c49
Docstring changes
