### 2019-11-08 samuelduchesne/archetypal#43 @samuelduchesne
Hi @brynpickering, you are right that E+ version handling is more or less patchy at the moment. It has been a goal for us to try to handle as much as possible version compatibility issues. This is why, for instance, archetypal tries to upgrade the idf file (if it is older than any installed E+ versions) before launching E+.

Then again, it makes me wonder what is the best path forward. Should `eppy` handle versions of energy plus natively? Or should `archetypal` continue to patch such functionalities? I would be interested in having your thoughts on that.

Anyhow, let us add a more explicit error handling and provide advice on how to solve them, as you suggest. For example, the `idflcass.idf_version_updater` does not handle well missing transitioning files at the moment, if I remember correctly; this should be addressed as well.

## Issue pathways
To organize this particular issue, I see the following use cases that need a closer look:
- [ ] User runs archetypal without any EnergyPlus installs
- [ ] User runs archetypal with an idf file older than the available install
- [ ] User runs archetypal with an idf file newer than the available install
- [ ] User runs archetypal with more than one version of E+ installed on their machine; which one should be used?
