### 2019-11-08 fd9a14a
Adds URLs to DOE commercial/residential reference buildings + code formating
