+++ paper/paper.bib
@@ -23,10 +23,10 @@
 }
 
 @misc{Bull2016,
-author = {Bull, Jamie},
-title = {{GeomEppy}},
-url = {https://github.com/jamiebull1/geomeppy},
-year = {2016}
+    author = {Bull, Jamie},
+    title = {{ GeomEppy }},
+    url = {https://github.com/jamiebull1/geomeppy},
+    year = {2016}
 }
 
 
@@ -65,18 +65,20 @@ year = {2016}
 }
 
 @misc{USDOE-BuildingEnergyCodesProgram2012,
-abstract = {These prototypes are then modified to accommodate four different heating system types and four foundation types typically found in new residential construction.},
-author = {{US DOE - Building Energy Codes Program}},
-booktitle = {US Department of Energy},
-title = {{Residential Prototype Building Models}},
-year = {2012}
+    abstract = {These prototypes are then modified to accommodate four different heating system types and four foundation types typically found in new residential construction.},
+    author = {{ US DOE - Building Energy Codes Program }},
+    booktitle = {US Department of Energy},
+    title = {{ Residential Prototype Building Models }},
+    url = {https://www.energycodes.gov/development/residential/iecc_models},
+    year = {2012}
 }
 
 @misc{USDOE-BuildingTechnologyOffice2018,
-address = {Washington},
-author = {{US DOE - Building Technology Office}},
-booktitle = {US Department of Energy},
-mendeley-groups = {Bs2019-LouisLeroy},
-title = {{Commercial Prototype Building Models}},
-year = {2018}
+    address = {Washington},
+    author = {{ US DOE - Building Technology Office }},
+    booktitle = {US Department of Energy},
+    mendeley-groups = {Bs2019-LouisLeroy},
+    title = {{ Commercial Prototype Building Models }},
+    url = {https://www.energycodes.gov/development/commercial/prototype_models},
+    year = {2018}
 }
