### 2019-11-08 #1833 @samuelduchesne
Hi @brynpickering and @ThibaultMarzullo! Sorry for the late reply on my end. I have been quite busy with the submission of my thesis this week. Now it is done! so I can commit to this review process! 💪

You have provided valuable comments and uncovered issues we weren't able to pick even with our CI integration. If you do not mind letting us some time to get over all the issues, we will try to address them as soon as possible.

For the issue of TRNSYS, how about we show a proof of use as a screencast video maybe? @louisleroy5, do you have some thoughts? Could be included in the documentation 🤔
