### 2019-11-08 samuelduchesne/archetypal#47 @samuelduchesne
Easy enough 😉

Added links to the references themselves. Is that what you meant @brynpickering ?

Should I close this issue now as other paper-related issues could be handled in issue #41
