### 2019-11-08 7acde2e
Makes sure internal gains are passed as kJ/h-m2 in the T3D file
