+++ archetypal/trnsys.py
@@ -2136,7 +2136,7 @@ def _write_equipment_gain(equipments, gainNum, lines, res, old_new_names):
         # Write gain name in lines
         lines.insert(gainNum + 1, "GAIN " + equipment.Name + "\n")
         areaMethod = "AREA_RELATED"
-        power = gain["Equipment/Floor Area {W/m2}"].values[0]
+        power = gain["Equipment/Floor Area {W/m2}"].values[0] / 1000 * 3600 # kJ/h-m2
         radFract = gain["Fraction Radiant"].values[0]
         lines.insert(
             gainNum + 2,
@@ -2169,7 +2169,7 @@ def _write_light_gain(gainNum, lights, lines, res, old_new_names):
         # Write gain name in lines
         lines.insert(gainNum + 1, "GAIN " + light.Name + "\n")
         areaMethod = "AREA_RELATED"
-        power = gain["Lights/Floor Area {W/m2}"].values[0]
+        power = gain["Lights/Floor Area {W/m2}"].values[0] / 1000 * 3600 # kJ/h-m2
         radFract = gain["Fraction Radiant"].values[0]
         lines.insert(
             gainNum + 2,
@@ -2200,7 +2200,7 @@ def _write_people_gain(gainNum, lines, peoples, res, old_new_names):
         # Write gain name in lines
         lines.insert(gainNum + 1, "GAIN " + people.Name + "\n")
         areaMethod = "AREA_RELATED"
-        power = gain["People/Floor Area {person/m2}"].values[0]
+        power = gain["People/Floor Area {person/m2}"].values[0] * 270 # kJ/h-m2
         radFract = gain["Fraction Radiant"].values[0]
         lines.insert(
             gainNum + 2,
