### 2019-11-11 6a41646
Adds fframe and uframe as kwargs input to characterize the window
