### 2019-11-11 7b2e229
New: Using heating and cooling power from IDF instead of infinite capacity
