### 2019-11-11 2c45d0e
Adds namespace registering function + example in docs
