### 2019-11-11 6aaa567
Improves dependencies of version 9.x for conversion from older IDF files
