### 2019-11-11 53aac9a
New: Gets TGround from IDF and use it as schedule in TRNBuild
