### 2019-11-11 415ebb8
Fix: makes sure array as float64 before rounding it
