### 2019-11-11 bd1b5a0
New: Uses sizing factor for heating and cooling systems from IDF
