### 2019-11-13 d49763b
adds a method to check compatibility of E+ versions on package import
