+++ archetypal/__init__.py
@@ -12,8 +12,10 @@ ep_version = "8-9-0"
 
 # warn if a newer version of archetypal is available
 from outdated import warn_if_outdated
+from .utils import warn_if_not_compatible
 
 warn_if_outdated("archetypal", __version__)
+warn_if_not_compatible()
 
 from .utils import *
 from .simple_glazing import *
