### 2019-11-13 b653efb
Adds docs in methodology section to explain the "T3D to BUI" section better
