### 2019-11-13 efdec4b
Fixes an issue with file paths in the upgrade transition process
