### 2019-11-13 7f023d2
Fix: Deepcopy file instead of copy
