### 2019-11-13 ae8191f
Adds new tests for coverage
