### 2019-11-13 bbd35f5
Renames function
