### 2019-11-14 f814e74
Adds 'fframe' and 'uframe' kwargs in cli and trnsys example
