### 2019-11-14 343ce15
Adds docs to explain new features of the IDF to TRNBuild converter
