### 2019-11-14 88123a4
Adds window fframe and uframe to write in lines
