+++ archetypal/cli.py
@@ -204,7 +204,7 @@ def cli(
     "--window",
     nargs=4,
     type=float,
-    default=(2.2, 0.64, 0.8, 0.05),
+    default=(2.2, 0.64, 0.8, 0.05, 0.15, 8.17),
     help="Specify window properties <u_value> <shgc> <t_vis> "
     "<tolerance>. Default = 2.2 0.64 0.8 0.05",
 )
