### 2019-11-14 ed4d795
Asserts all files needed exist before running EP simulation
