### 2019-11-14 028c4a1
New test for cli convert()
