### 2019-11-14 10a0b47
Adds infos in "window" help
