### 2019-11-14 3fa0dbc
Adds info on default values used for cli arguments
