+++ docs/converter_bui.rst
@@ -78,7 +78,7 @@ enclosed in quotation marks.
 enclosed in quotation marks.
 
 3. ``OUTPUT_FOLDER`` is the folder where we want the output folders to be written. If there are space characters in
-the path, it should enclosed in quotation marks.
+the path, it should enclosed in quotation marks. If nothing is passed, the output folder will be the current working directory.
 
 Here is an example. Make sure to replace the last two arguments with the idf file path and the output folder path
 respectively.
@@ -98,20 +98,22 @@ Users can chose to return a combination of flags
 
         archetypal convert -i -t -d "/Users/Documents/NECB 2011 - Warehouse.idf" "/Users/Documents/WIP"
 
-    - ``--window-lib`` is the path of the window library (W74-lib.dat).
-        This library must be in the same format as the Berkeley Lab library used by default in TRNBuild.
+    - ``--window-lib`` is the path of the window library (W74-lib.dat). This library must be in the same format as the Berkeley Lab library used by default in TRNBuild.
+        If nothing is passed, the "W74-lib.dat" file available in the package "ressources" folder will be used.
 
     .. code-block:: python
 
         archetypal convert --window-lib "/Users/Documents/W74-lib.dat" "/Users/Documents/NECB 2011 - Warehouse.idf" "/Users/Documents/WIP"
 
     - ``--trnsidf-exe`` is the path of the trnsidf.exe executable. Usually located in the TRNSYS18 folder under "Building/trnsIDF/trnsidf.exe".
+        If nothing is passed, the following path will be used : "C:TRNSYS18\\Building\\trnsIDF\\trnsidf.exe".
 
     .. code-block:: python
 
         archetypal convert --trnsidf-exe "C:TRNSYS18\\Building\\trnsIDF\\trnsidf.exe" "/Users/Documents/NECB 2011 - Warehouse.idf" "/Users/Documents/WIP"
 
     - ``--template`` is the path of the .d18 template file (usually in the same directory as the `trnsidf.exe` executable).
+        If nothing is passed, the following path will be used : "C:TRNSYS18\\Building\\trnsIDF\\NewFileTemplate.d18".
 
     .. code-block:: python
 
@@ -123,11 +125,12 @@ Users can chose to return a combination of flags
 
         archetypal convert --log-clear-names "/Users/Documents/NECB 2011 - Warehouse.idf" "/Users/Documents/WIP"
 
-    - ``--window`` specifies the window properties <u_value> <shgc> <t_vis> <tolerance> <fframe> <uframe>
+    - ``--window`` specifies the window properties <u_value (W/m²-K)> <shgc (-)> <t_vis (-)> <tolerance (-)> <fframe (-)> <uframe (kJ/m²-K-h)>.
+        If nothing is passed, the following values will be used : 2.2 0.65 0.8 0.05 0.15 8.17
 
     .. code-block:: python
 
-        archetypal convert --window 2.2 0.65 0.8 0.05 0.15 5.4 "/Users/Documents/NECB 2011 - Warehouse.idf" "/Users/Documents/WIP"
+        archetypal convert --window 2.2 0.65 0.8 0.05 0.15 8.17 "/Users/Documents/NECB 2011 - Warehouse.idf" "/Users/Documents/WIP"
 
     - ``--ordered`` sorts the idf object names
 
