### 2019-11-15 a50634f
Corrects typo in cli
