### 2019-11-15 e1fc579
Fix: Change nargs to 6 for window option
