### 2019-11-15 655981d
Parametric test to test all the cli possibilities for convert method
