+++ tests/test_cli.py
@@ -1,3 +1,4 @@
+import pytest
 from click.testing import CliRunner
 
 from archetypal import get_eplus_dire
@@ -8,20 +9,10 @@ from path import Path
 class TestCli:
     """Defines tests for usage of the archetypal Command Line Interface"""
 
-    def test_convert(self, config):
-        """Tests the 'reduce' method"""
-        runner = CliRunner()
-        examples = get_eplus_dire() / "ExampleFiles"
-        necb = Path("tests/input_data/necb")
-        test_file = "tests/input_data/trnsys/5ZoneGeometryTransform.idf"
-        test_file_list = [
-            "tests/input_data/umi_samples/B_Off_0.idf",
-            "tests/input_data/umi_samples/B_Res_0_Masonry.idf",
-        ]
-        test_files = necb.glob("*Retail*.idf")
-        result = runner.invoke(
-            cli,
+    @pytest.fixture(
+        params=[
             [
+                1,
                 "--use-cache",
                 "--cache-folder",
                 "tests/.temp/cache",
@@ -33,13 +24,209 @@ class TestCli:
                 "tests/.temp/logs",
                 "--log-console",
                 "convert",
-                test_file,
+                "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
                 "--trnsidf-exe",
                 "docker/trnsidf/trnsidf.exe",
             ],
-            catch_exceptions=False,
-        )
+            [
+                2,
+                "--use-cache",
+                "--cache-folder",
+                "tests/.temp/cache",
+                "--data-folder",
+                "tests/.temp/data",
+                "--imgs-folder",
+                "tests/.temp/images",
+                "--logs-folder",
+                "tests/.temp/logs",
+                "--log-console",
+                "convert",
+                "-i",
+                "-t",
+                "-d",
+                "tests/input_data/trnsys/simple_2_zone.idf",
+                "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
+                "--trnsidf-exe",
+                "docker/trnsidf/trnsidf.exe",
+            ],
+            [
+                3,
+                "--use-cache",
+                "--cache-folder",
+                "tests/.temp/cache",
+                "--data-folder",
+                "tests/.temp/data",
+                "--imgs-folder",
+                "tests/.temp/images",
+                "--logs-folder",
+                "tests/.temp/logs",
+                "--log-console",
+                "convert",
+                "--window-lib",
+                "archetypal/ressources/W74-lib.dat",
+                "tests/input_data/trnsys/simple_2_zone.idf",
+                "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
+                "--trnsidf-exe",
+                "docker/trnsidf/trnsidf.exe",
+            ],
+            [
+                4,
+                "--use-cache",
+                "--cache-folder",
+                "tests/.temp/cache",
+                "--data-folder",
+                "tests/.temp/data",
+                "--imgs-folder",
+                "tests/.temp/images",
+                "--logs-folder",
+                "tests/.temp/logs",
+                "--log-console",
+                "convert",
+                "--template",
+                "archetypal/ressources/NewFileTemplate.d18",
+                "tests/input_data/trnsys/simple_2_zone.idf",
+                "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
+                "--trnsidf-exe",
+                "docker/trnsidf/trnsidf.exe",
+            ],
+            [
+                5,
+                "--use-cache",
+                "--cache-folder",
+                "tests/.temp/cache",
[…]
