### 2019-11-15 1bea640
Creates reference docs for dataportal, energydataframe and energyseries
