### 2019-11-15 c0571de
Moves generic function in utils
