### 2019-11-15 550d097
Renames functions
