### 2019-11-15 2f53ab1
Adds docstring
