### 2019-11-15 05ec25b
Adds/cleans docstring
