### 2019-11-15 f958ef0
Adds reference for DenergySeries, ReportData, Tabular Data, TRNSYS, UmiTemplate and Utils
