+++ docs/package_modules.rst
@@ -129,8 +129,8 @@ EnergyDataFrame
     :nosignatures:
     :toctree: reference/
 
-    set_unit
-    discretize_tsam
+    EnergyDataFrame.set_unit
+    EnergyDataFrame.discretize_tsam
     plot_energydataframe_map
 
 
@@ -144,6 +144,125 @@ EnergySeries
     :nosignatures:
     :toctree: reference/
 
+    EnergySeries.from_sqlite
+    EnergySeries.unit_conversion
+    EnergySeries.concurrent_sort
+    EnergySeries.normalize
+    EnergySeries.ldc_source
+    EnergySeries.source_side
+    EnergySeries.discretize_tsam
+    EnergySeries.discretize
+    EnergySeries.plot3d
+    EnergySeries.plot2d
+    EnergySeries.p_max
+    EnergySeries.p_max
+    EnergySeries.monthly
+    EnergySeries.capacity_factor
+    EnergySeries.bin_edges
+    EnergySeries.time_at_min
+    EnergySeries.bin_scaling_factors
+    EnergySeries.duration_scaling_factor
+    EnergySeries.ldc
+    EnergySeries.nseries
     save_and_show
     plot_energyseries
     plot_energyseries_map
+
+
+Report Data
+-----------
+
+.. currentmodule:: archetypal.reportdata
+
+.. autosummary::
+    :template: autosummary.rst
+    :nosignatures:
+    :toctree: reference/
+
+    ReportData.from_sql_dict
+    ReportData.from_sqlite
+    ReportData.heating_load
+    ReportData.filter_report_data
+    ReportData.sorted_values
+
+
+Tabular Data
+------------
+
+.. currentmodule:: archetypal.tabulardata
+
+.. autosummary::
+    :template: autosummary.rst
+    :nosignatures:
+    :toctree: reference/
+
+    TabularData.from_sql
+    TabularData.filter_tabular_data
+
+IDF to BUI module
+-----------------
+
+.. currentmodule:: archetypal.trnsys
+
+.. autosummary::
+    :template: autosummary.rst
+    :nosignatures:
+    :toctree: reference/
+
+    convert_idf_to_trnbuild
+    get_idf_objects
+    clear_name_idf_objects
+    zone_origin
+    closest_coords
+    parse_window_lib
+    choose_window
+    trnbuild_idf
+
+
+UMI Template
+------------
+
+.. currentmodule:: archetypal.umi_template
+
+.. autosummary::
+    :template: autosummary.rst
+    :nosignatures:
+    :toctree: reference/
+
+    UmiTemplate.from_idf
+    UmiTemplate.from_json
+    UmiTemplate.to_json
+
+
+Utils
+-----
+
+.. currentmodule:: archetypal.utils
+
+.. autosummary::
+    :template: autosummary.rst
+    :nosignatures:
+    :toctree: reference/
+
+    validate_trnsys_folder
+    log
+    load_umi_template_objects
+    umi_template_object_to_dataframe
+    get_list_of_common_umi_objects
+    newrange
+    date_transform
+    weighted_mean
+    top
+    copy_file
+    piecewise
+    rmse
+    checkStr
+    write_lines
+    load_umi_template
+    check_unique_name
+    angle
+    float_round
+    timeit
+    lcm
+    recursive_len
+    rotate
