### 2019-11-18 f211850
Fixes an issue where the idf file transition would occur to the latest installed version instead of using the package global ep_version
This is more consistent with normal behavior where a user can define the version of EnergyPlus to use in config settings
