### 2019-11-18 69a44c9
Moves ep_version to settings module and adds it to the config method
