+++ tests/conftest.py
@@ -60,13 +60,13 @@ def idf_source(request):
 @pytest.fixture(scope="session")
 def config():
     ar.config(
-        log_console=True,
-        log_file=True,
-        use_cache=True,
         data_folder="tests/.temp/data",
         logs_folder="tests/.temp/logs",
         imgs_folder="tests/.temp/imgs",
         cache_folder="tests/.temp/cache",
+        use_cache=True,
+        log_file=True,
+        log_console=True,
         umitemplate="tests/input_data/umi_samples" "/BostonTemplateLibrary_2.json",
     )
 
