+++ archetypal/__init__.py
@@ -7,8 +7,6 @@
 
 # Version of the package
 __version__ = "1.3.0"
-# Latest version of EnergyPlus compatible with archetypal
-ep_version = "8-9-0"
 
 # warn if a newer version of archetypal is available
 from outdated import warn_if_outdated
