### 2019-11-18 e5fc7a5
Code cleanup
