+++ archetypal/cli.py
@@ -7,14 +7,25 @@
 import os
 from collections import defaultdict
 
-import archetypal
 import click
-from archetypal import settings, cd, load_idf
+
+from archetypal import (
+    settings,
+    cd,
+    load_idf,
+    convert_idf_to_trnbuild,
+    get_eplus_dirs,
+    parallel_process,
+    run_eplus,
+    IDF,
+    UmiTemplate,
+    config,
+)
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 
 
-class Config(object):
+class CliConfig(object):
     def __init__(self):
         self.data_folder = settings.data_folder
         self.logs_folder = settings.logs_folder
@@ -31,7 +42,7 @@ class Config(object):
         self.trnsys_default_folder = settings.trnsys_default_folder
 
 
-pass_config = click.make_pass_decorator(Config, ensure=True)
+pass_config = click.make_pass_decorator(CliConfig, ensure=True)
 
 
 @click.group(context_settings=CONTEXT_SETTINGS)
@@ -98,7 +109,7 @@ pass_config = click.make_pass_decorator(Config, ensure=True)
 )
 @pass_config
 def cli(
-    config,
+    cli_config,
     data_folder,
     logs_folder,
     imgs_folder,
@@ -117,31 +128,35 @@ def cli(
     Visit archetypal.readthedocs.io for the online documentation
 
     Args:
-        config:
-        data_folder:
-        logs_folder:
-        imgs_folder:
-        cache_folder:
-        use_cache:
-        log_file:
-        log_console:
-        log_level:
-        log_name:
-        log_filename:
-        trnsys_default_folder:
+        cli_config: Package configuration. See :class:`CliConfig` for more details.
+        data_folder (str): where to save and load data files.
+        logs_folder (str): where to write the log files.
+        imgs_folder (str): where to save figures.
+        cache_folder (str): where to save the simluation results.
+        use_cache (bool): if True, use a local cache to save/retrieve many of
+            archetypal outputs such as EnergyPlus simulation results. This can
+            save a lot of time by not calling the simulation and dataportal APIs
+            repetitively for the same requests.
+        log_file (bool): if true, save log output to a log file in logs_folder.
+        log_console (bool): if true, print log output to the console.
+        log_level (int): one of the logger.level constants.
+        log_name (str): name of the logger.
+        log_filename (str): name of the log file.
+        trnsys_default_folder (str): root folder of TRNSYS install.
     """
-    config.data_folder = data_folder
-    config.logs_folder = logs_folder
-    config.imgs_folder = imgs_folder
-    config.cache_folder = cache_folder
-    config.use_cache = use_cache
-    config.log_file = log_file
-    config.log_console = log_console
-    config.log_level = log_level
-    config.log_name = log_name
-    config.log_filename = log_filename
-    config.trnsys_default_folder = trnsys_default_folder
-    archetypal.config(**config.__dict__)
+    cli_config.data_folder = data_folder
+    cli_config.logs_folder = logs_folder
+    cli_config.imgs_folder = imgs_folder
+    cli_config.cache_folder = cache_folder
+    cli_config.use_cache = use_cache
+    cli_config.log_file = log_file
+    cli_config.log_console = log_console
+    cli_config.log_level = log_level
+    cli_config.log_name = log_name
+    cli_config.log_filename = log_filename
+    cli_config.trnsys_default_folder = trnsys_default_folder
+    # apply new config params
+    config(**cli_config.__dict__)
 
 
 @cli.command()
@@ -279,7 +294,7 @@ def convert(
         "tolerance": tolerance,
     }
     with cd(output_folder):
[…]
