### 2019-11-18 0c454e8
Attempt to fix plot_graph2d error in travis CI
