### 2019-11-18 e90f4e2
Moves ep_version to global cli settings, ie., before a command
