+++ tests/test_cli.py
@@ -12,14 +12,10 @@ class TestCli:
     def test_reduce(self, config):
         """Tests the 'reduce' method"""
         runner = CliRunner()
-        examples = get_eplus_dirs(ep_version) / "ExampleFiles"
-        necb = Path("tests/input_data/necb")
-        test_file = examples / "2ZoneDataCenterHVAC_wEconomizer.idf"
         test_file_list = [
             "tests/input_data/umi_samples/B_Off_0.idf",
             "tests/input_data/umi_samples/B_Res_0_Masonry.idf",
         ]
-        test_files = necb.glob("*Retail*.idf")
         result = runner.invoke(
             cli,
             [
@@ -33,6 +29,8 @@ class TestCli:
                 "--logs-folder",
                 "tests/.temp/logs",
                 "--log-console",
+                "--ep_version",
+                "8-9-0",
                 "reduce",
                 "-n",
                 "Retail",
