### 2019-11-19 34cefb1
Adds an option where schedule can be used as input or schedules
The user can choose if the yearly schedules are used as INPUTS , meaning that he will have to link those inputs with the output schedule csv file. Otherwise he can choose to directly use the yearly schedules as SCHEDULES, but this last method can make TRNBuild crash if too many/too long schedules
