+++ archetypal/trnsys.py
@@ -47,6 +47,7 @@ def convert_idf_to_trnbuild(
     trnsidf_exe=None,
     template=None,
     log_clear_names=False,
+    schedule_as_input=True,
     **kwargs
 ):
     """Convert regular IDF file (EnergyPlus) to TRNBuild file (TRNSYS)
@@ -237,7 +238,14 @@ def convert_idf_to_trnbuild(
 
     # Writing zones in lines
     win_slope_dict = _write_zone_buildingSurf_fenestrationSurf(
-        buildingSurfs, coordSys, fenestrationSurfs, idf_2, lines, n_ground, zones
+        buildingSurfs,
+        coordSys,
+        fenestrationSurfs,
+        idf_2,
+        lines,
+        n_ground,
+        zones,
+        schedule_as_input,
     )
     # endregion
 
@@ -261,7 +269,9 @@ def convert_idf_to_trnbuild(
     # endregion
 
     # region Write SCHEDULES from IDF to lines (T3D)
-    schedules_not_written = _write_schedules(lines, schedule_names, schedules)
+    schedules_not_written = _write_schedules(
+        lines, schedule_names, schedules, schedule_as_input, idf_file
+    )
     # endregion
 
     # region Write WINDOWS chosen by the user (from Berkeley lab library) in
@@ -361,6 +371,7 @@ def convert_idf_to_trnbuild(
         schedules_not_written,
         res,
         old_new_names,
+        schedule_as_input,
     )
 
     # Save B18 file at output_folder
@@ -413,6 +424,7 @@ def gains_to_b18(
     schedules_not_written,
     res,
     old_new_names,
+    schedule_as_input,
 ):
     peoples_in_zone = zone_where_gain_is(peoples, zones, zonelists)
     lights_in_zone = zone_where_gain_is(lights, zones, zonelists)
@@ -429,6 +441,7 @@ def gains_to_b18(
             res,
             old_new_names,
             "People",
+            schedule_as_input,
         )
         # Write light gains
         _write_gain_to_b18(
@@ -440,6 +453,7 @@ def gains_to_b18(
             res,
             old_new_names,
             "Lights",
+            schedule_as_input,
         )
         # Write equipment gains
         _write_gain_to_b18(
@@ -451,6 +465,7 @@ def gains_to_b18(
             res,
             old_new_names,
             "ElectricEquipment",
+            schedule_as_input,
         )
 
 
@@ -463,6 +478,7 @@ def _write_gain_to_b18(
     res,
     old_new_names,
     string,
+    schedule_as_input,
 ):
     for gain in gains:
         if zone.Name in gains_in_zone[gain.Name]:
@@ -479,15 +495,26 @@ def _write_gain_to_b18(
             if schedule in schedules_not_written:
                 continue
             # Write
-            b18_lines.insert(
-                regimeNum,
-                " GAIN= "
-                + gain.Name
-                + " : SCALE= SCHEDULE 1*"
-                + schedule
-                + " : GEOPOS=0 : SCALE2= 1 : FRAC_REFAREA= 1"
-                + "\n",
-            )
+            if schedule_as_input:
+                b18_lines.insert(
+                    regimeNum,
+                    " GAIN= "
+                    + gain.Name
+                    + " : SCALE= INPUT 1*"
+                    + schedule
+                    + " : GEOPOS=0 : SCALE2= 1 : FRAC_REFAREA= 1"
+                    + "\n",
+                )
+            else:
+                b18_lines.insert(
+                    regimeNum,
+                    " GAIN= "
+                    + gain.Name
+                    + " : SCALE= SCHEDULE 1*"
+                    + schedule
[…]
