### 2019-11-19 dc50b98
Adds docstring
