### 2019-11-19 c110853
Adds cli option to write schedules as INPUTS or SCHEDULES
