### 2019-11-19 99b1ffd
Adds option argument in cli for the schedule writing
