### 2019-11-19 68291ca
Adds test for cli "--schedule-as-input" option
