### 2019-11-19 ec5a5fd
Do not run convert tests on TRAVIS (trnsidf.exe not available)
