+++ tests/test_cli.py
@@ -1,4 +1,5 @@
 import pytest
+import os
 from click.testing import CliRunner
 
 from archetypal import get_eplus_dire
@@ -243,6 +244,10 @@ class TestCli:
 
         yield request.param[1:]
 
+    @pytest.mark.xfail(
+        "TRAVIS" in os.environ and os.environ["TRAVIS"] == "true",
+        reason="Skipping this test on Travis CI.",
+    )
     def test_convert(self, config, cli_args):
         """Tests the 'reduce' method"""
         runner = CliRunner()
