### 2019-11-19 26ff6f2
Optimizes imports for trnsys.py module
