### 2019-11-20 7c0d855
Better handling of EnergyPlus version errors when idf file version is different than specified EP version
Creates new EnergyPlusVersionError class
