### 2019-11-20 d166297
Better handling of EnergyPlus version errors when idf file version is different than specified EP version
Creates new EnergyPlusVersionError class
