### 2019-11-20 880cbcb
Adds ep_version as option in cli
