### 2019-11-20 19cd807
Returns all possible output reports in the IDF object
