### 2019-11-20 6e2c2ac
Insures a ReportData is retuned in classmethod from_sqlite
