### 2019-11-20 2fd055a
Adds setpoint schedules to schedules dict
