### 2019-11-20 ff5517e
Refactor variable name + runs black
