### 2019-11-20 b280ce0
Writes setpoints as schedules (8760 values) + refactor variable name (res to htm)
