+++ archetypal/trnsys.py
@@ -319,11 +319,13 @@ def convert_idf_to_trnbuild(
     # endregion
 
     # region Write GAINS (People, Lights, Equipment) from IDF to lines (T3D)
-    _write_gains(equipments, lights, lines, peoples, res, old_new_names)
+    _write_gains(equipments, lights, lines, peoples, htm, old_new_names)
     # endregion
 
     # region Write basic conditioning systems (HEATING and COOLING) from IDF to lines (T3D)
-    heat_name, cool_name = _write_conditioning(res, lines)
+    heat_dict, cool_dict = _write_conditioning(
+        htm, lines, schedules, old_new_names, schedule_as_input
+    )
     # endregion
 
     # region Write SCHEDULES from IDF to lines (T3D)
@@ -374,12 +376,12 @@ def convert_idf_to_trnbuild(
     # Run trnsidf to convert T3D to BUI
     log("Converting t3d file to bui file. Running trnsidf.exe...")
     dck = return_dck
-    nonum = kwargs.get("nonum", False)
-    N = kwargs.get("N", False)
-    geo_floor = kwargs.get("geo_floor", 0.6)
-    refarea = kwargs.get("refarea", False)
-    volume = kwargs.get("volume", False)
-    capacitance = kwargs.get("capacitance", False)
+    nonum = kwargs.pop("nonum", False)
+    N = kwargs.pop("N", False)
+    geo_floor = kwargs.pop("geo_floor", 0.6)
+    refarea = kwargs.pop("refarea", False)
+    volume = kwargs.pop("volume", False)
+    capacitance = kwargs.pop("capacitance", False)
     trnbuild_idf(
         t3d_path,
         output_folder=output_folder,
@@ -413,10 +415,10 @@ def convert_idf_to_trnbuild(
         b18_lines = b18_file.readlines()
 
     # Adds conditionning to B18 file
-    conditioning_to_b18(b18_lines, heat_name, cool_name, zones, old_new_names)
+    conditioning_to_b18(b18_lines, heat_dict, cool_dict, zones, old_new_names)
 
     # Adds infiltration to b18 file
-    infilt_to_b18(b18_lines, zones, res)
+    infilt_to_b18(b18_lines, zones, htm)
 
     # Adds internal gain to b18 file
     gains_to_b18(
@@ -575,21 +577,21 @@ def _write_gain_to_b18(
                 )
 
 
-def conditioning_to_b18(b18_lines, heat_name, cool_name, zones, old_new_names):
+def conditioning_to_b18(b18_lines, heat_dict, cool_dict, zones, old_new_names):
     for zone in zones:
         # Heating
-        _write_heat_cool_to_b18(heat_name, old_new_names, zone, b18_lines, " HEATING")
+        _write_heat_cool_to_b18(heat_dict, old_new_names, zone, b18_lines, " HEATING")
         # Cooling
-        _write_heat_cool_to_b18(cool_name, old_new_names, zone, b18_lines, " COOLING")
+        _write_heat_cool_to_b18(cool_dict, old_new_names, zone, b18_lines, " COOLING")
 
 
-def _write_heat_cool_to_b18(list_name, old_new_names, zone, b18_lines, string):
-    for key in list_name.keys():
+def _write_heat_cool_to_b18(list_dict, old_new_names, zone, b18_lines, string):
+    for key in list_dict.keys():
         if old_new_names[zone.Name.upper()][0] in key:
             f_count = checkStr(b18_lines, "Z o n e  " + zone.Name)
             regimeNum = checkStr(b18_lines, "REGIME", f_count)
             # Write
-            b18_lines.insert(regimeNum, string + " = " + list_name[key] + "\n")
[…]
