### 2019-11-20 957c2e5
Factorize code
