### 2019-11-20 5d60afb
Refactor tests to work
