### 2019-11-21 dcc9bc6
Deletes corrupt IDF files
