+++ tests/test_trnsys.py
@@ -548,7 +548,6 @@ def test_trnbuild_from_idf_parallel(config, trnbuild_file):
         "NECB 2011 - HighRise Apartment.idf",
         "NECB 2011 - Hospital.idf",
         "NECB 2011 - Large Hotel.idf",
-        "NECB 2011 - Medium Office.idf",
         "NECB 2011 - MidRise Apartment.idf",
         "NECB 2011 - Outpatient.idf",
         "NECB 2011 - Primary School.idf",
@@ -611,7 +610,6 @@ def test_trnbuild_from_idf_parallel_darwin_or_linux(config):
         "NECB 2011 - HighRise Apartment.idf",
         "NECB 2011 - Hospital.idf",
         "NECB 2011 - Large Hotel.idf",
-        "NECB 2011 - Medium Office.idf",
         "NECB 2011 - MidRise Apartment.idf",
         "NECB 2011 - Outpatient.idf",
         "NECB 2011 - Primary School.idf",
