### 2019-11-21 a6cebd7
Adds new kwargs in test
