+++ docs/troubleshooting.rst
@@ -1,0 +1,19 @@
+Troubleshooting
+===============
+
+MacOS Catalina
+--------------
+
+
+
+Missing transition files
+------------------------
+
+1. Go on the Knowledgebase_ of EnergyPlus to download the transition files
+    depending on the platform you are using
+2. Extract the files to EnergyPlus installation path, in PreProcess/IDFVersionUpdater/
+    (e.g. /Applications/EnergyPlus-n-n-n/PreProcess/IDFVersionUpdater/ for MacOS).
+
+.. image:: images/extract_transition_EP_files.gif
+
+.. _Knowledgebase: http://energyplus.helpserve.com/Knowledgebase/List/Index/46/converting-older-version-files
