### 2019-11-21 5b3756c
Adds ability to specify the E+ version to use when loading an IDF file
