### 2019-11-21 1b42154
Raises error if transition file not found - TODO: add link to doc
