### 2019-11-21 a19e30f
Updates missing transition programs with link to document + new troubleshooting documentation
