### 2019-11-22 add852a
Fix: There is no "weather_file" input in trnbuild_idf()
