### 2019-11-22 a4574e1
Makes sure INPUTS line not too long when adding the schedules
