### 2019-11-22 dbc2d5c
Fix increment value (13 instead of 14) + renames variable
