### 2019-11-22 548c4f8
Checks if infiltration exists before setting a value
