### 2019-11-22 900ac62
Schedule_as_input=False for simple idf
