+++ tests/test_trnsys.py
@@ -742,5 +742,6 @@ def test_trnbuild_from_simple_idf(config):
         window_lib=window_filepath,
         template="tests/input_data/trnsys/NewFileTemplate.d18",
         trnsidf_exe="docker/trnsidf/trnsidf.exe",
+        schedule_as_input=False,
         **kwargs_dict
     )
