### 2019-11-25 #1833 @kyleniemeyer
Hi @samuelduchesne, just wanted to check in and see how your work was going.
