### 2019-11-25 samuelduchesne/archetypal#47 @brynpickering
Yes, happy for this to be closed. Although. I would first check the most recently built paper PDF, there seems to be some issue with the USDOE reference in the main text.
