### 2019-11-25 samuelduchesne/archetypal#43 @brynpickering
PR looks like it addresses most of the concerns I raised, and the use cases you mention in your comment are those that I agree should be addressed. In addition I would include 'User runs archetypal with an idf file older than the available install, but available install is newer than we have accounted for in our conversion scripts' (i.e. the case where I tried to use v9.2.0 before you updated the code with #49). Also, could you have a test that runs for each of these use cases as part of your continuous integration?
