### 2019-11-25 samuelduchesne/archetypal#43 @samuelduchesne
Yes, It's possible to test other EnergyPlus versions in the CI. Let me see how I can implement this. I'll add your suggestion to the series of tests.
