### 2019-11-25 ad909b6
Fixes an issue with the rendering of USDOE refs.
