+++ paper/paper.md
@@ -77,7 +77,7 @@ Since it can be long and error-prone to create archetype buildings by hand, conv
 EnergyPlus models to TrnBuild models emerged as a way of speeding both the intermodel
 comparisons and the supplemental model creation. That is to say, a large repository of
 prototype building models exists in the literature with a large majority developed in the
-popular EnergyPlus environment [@USDOE2012, @USDOE2018]. `archetypal ` answers a need from
+popular EnergyPlus environment [@USDOE2012; @USDOE2018]. `archetypal ` answers a need from
 researchers and building energy model specialists to create TrnBuild Models from existing
 EnergyPlus models.
 
