### 2019-11-25 samuelduchesne/archetypal#47 @samuelduchesne
Good catch! Latest PDF looks good now.
