### 2019-11-28 a828ade
Adds McTavish weather file
