### 2019-11-28 8462329
Set Solar_Abs of construct with Outside layer
