### 2019-11-29 3021c92
Fix code to avoid the use of "try...except"
