+++ archetypal/trnsys.py
@@ -2621,7 +2621,7 @@ def _write_constructions(constr_list, idf, lines, mat_name, materials):
         except:
             condition = False
         if condition:
-            lines.insert(constructionNum + 6, "!- HFRONT   = 11 : HBACK= 0.002\n")
+            lines.insert(constructionNum + 6, "!- HFRONT   = 11 : HBACK= 0.0005\n")
         else:
             lines.insert(constructionNum + 6, "!- HFRONT   = 11 : HBACK= 64\n")
 
