### 2019-11-29 6b51383
New: Writes initial zone temperature in B18
