### 2019-12-02 930e59f
Enables CI testing on MacOs platform
