+++ .travis.yml
@@ -1,6 +1,8 @@
 dist: xenial
 language: python
-os: linux
+os: 
+  - linux
+  - osx
 env: ARCHETYPAL_INTEGRATION=True ENERGYPLUS_VERSION=8.9.0 ENERGYPLUS_SHA=40101eaafd ENERGYPLUS_INSTALL_VERSION=8-9-0
 
 addons:
