### 2019-12-02 0865c99
Uses matrix.include to sperate linux and osx behaviors
