+++ .travis.yml
@@ -63,7 +63,8 @@ before_install:
     echo "y\r" | sudo ./$ENERGYPLUS_DOWNLOAD_FILENAME.$EXT; fi
   
   - sudo rm $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT
-  - sudo rm 8230
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
+    sudo rm 8230; fi
   
   # Install python
   - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
