### 2019-12-02 0f67ae4
Adds python installation + moves install/script out of matrix
