### 2019-12-02 c3cf1b9
Adds missing "fi" at end of condition block
