### 2019-12-02 fbe4cfd
Fixes install of python on linux
- adds missing semicolon
- changes install from setup.py to pip install .[dev]
