### 2019-12-02 64cdf91
Adds extra conversion files for MacOs build
