### 2019-12-03 68afb83
Cleans downloaded files + removes problematic command on osx
