### 2019-12-03 de4b6de
Enhances the way to get and write solar absorptance
