### 2019-12-03 f885cbb
Adds windows extras info (not used)
