### 2019-12-04 0050141
Skips Trnsys test on two different MacOs cases
1-The trnsidf.exe is not specified in docker/trnsidf path
2-The MacOs version is higher than 10.15

Since Catalina, 32 bit apps can't be run, even using wine.
