### 2019-12-04 3d2dba9
Proper error message when TRNSYS path not available
