### 2019-12-22 #1833 @samuelduchesne
@labarba — still a few things to address. With my thesis defense just last week, things were pretty hectic. I’ll finish the review right after the holidays. Sorry if it takes so long; I want this to be right.
