### 2020-01-06 395c49a
Streamlined bahvior for sql and sql_file properties
