### 2020-01-06 cd16b9d
Adds the weather_file path in the loading method inside run_eplus()
