### 2020-01-06 7fe15ca
Fixes an issue where the column name is different in Epversions > 8-9-0
Solution is to use partial string matching with "filter" than apply a squeeze() to return a series
