### 2020-01-06 84211f6
Fixes schedule order + uses calendar modules instead of an array
