+++ tests/test_schedules.py
@@ -1,6 +1,7 @@
 import os
 
 import pytest
+
 from archetypal import Schedule, load_idf, copy_file, run_eplus, UmiSchedule, config
 
 
