### 2020-01-06 f99bd4e
Fixes schedule order + uses calendar modules instead of an array (2)
