### 2020-01-06 81f7545
Improves the way Nat Vent is detected
