### 2020-01-13 a37907b
sets instances of E+ version number to package-wide variable + updates strings
