### 2020-01-13 e9146d3
Fixes an issue where the zone was not in the nat_df
