### 2020-01-13 bdc3c17
Updates travis to use version 9.2
