### 2020-01-14 000c9b0
Better default cli options
Less arguments are needed to perform a model reduction
