+++ archetypal/cli.py
@@ -6,6 +6,9 @@
 ################################################################################
 import os
 from collections import defaultdict
+from typing import Any, Union
+
+from path import Path
 
 import click
 
@@ -20,6 +23,7 @@ from archetypal import (
     IDF,
     UmiTemplate,
     config,
+    log,
 )
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
@@ -72,24 +76,24 @@ pass_config = click.make_pass_decorator(CliConfig, ensure=True)
     default=settings.cache_folder,
 )
 @click.option(
-    "--use-cache",
-    is_flag=True,
+    "-c/-nc",
+    "--use-cache/--no-cache",
     help="Use a local cache to save/retrieve many of "
     "archetypal outputs such as EnergyPlus simulation results",
-    default=settings.use_cache,
+    default=True,
 )
 @click.option(
+    "-l",
     "--log-file",
     is_flag=True,
     help="save log output to a log file in logs_folder",
     default=settings.log_file,
 )
 @click.option(
-    "--log-console",
-    "--verbose",
-    "-v",
-    is_flag=True,
-    default=settings.log_console,
+    "--verbose/--noverbose",
+    "-v/-nv",
+    "log_console",
+    default=True,
     help="print log output to the console",
 )
 @click.option(
@@ -133,25 +137,7 @@ def cli(
     """archetypal: Retrieve, construct, simulate, convert and analyse building
     simulation templates
 
-    Visit archetypal.readthedocs.io for the online documentation
-
-    Args:
-        cli_config: Package configuration. See :class:`CliConfig` for more details.
-        data_folder (str): where to save and load data files.
-        logs_folder (str): where to write the log files.
-        imgs_folder (str): where to save figures.
-        cache_folder (str): where to save the simluation results.
-        use_cache (bool): if True, use a local cache to save/retrieve many of
-            archetypal outputs such as EnergyPlus simulation results. This can
-            save a lot of time by not calling the simulation and dataportal APIs
-            repetitively for the same requests.
-        log_file (bool): if true, save log output to a log file in logs_folder.
-        log_console (bool): if true, print log output to the console.
-        log_level (int): one of the logger.level constants.
-        log_name (str): name of the logger.
-        log_filename (str): name of the log file.
-        trnsys_default_folder (str): root folder of TRNSYS install.
-        ep_version (str): EnergyPlus version to use. eg.: "9-2-0".
+    Visit archetypal.readthedocs.io for the online documentation.
     """
     cli_config.data_folder = data_folder
     cli_config.logs_folder = logs_folder
@@ -276,25 +262,6 @@ def convert(
 ):
     """Convert regular IDF file (EnergyPlus) to TRNBuild file (TRNSYS) The
     output folder path defaults to the working directory. Equivalent to '.'
-
-    Args:
-        idf_file:
-        window_lib:
-        return_idf:
-        return_t3d:
-        return_dck:
-        output_folder:
-        trnsidf_exe:
-        template:
-        log_clear_names:
-        window:
-        ordered:
-        nonum:
-        batchjob:
-        geofloor:
-        refarea:
-        volume:
-        capacitance:
     """
     u_value, shgc, t_vis, tolerance = window
     window_kwds = {
@@ -341,37 +308,33 @@ def convert(
 
 @cli.command()
 @click.argument("idf", nargs=-1)
-@click.option(
-    "--name",
-    "-n",
-    type=click.STRING,
-    default="umitemplate",
-    help="The name of the output json file",
-)
+@click.argument("output", type=click.Path(), default="myumitemplate.json")
 @click.option(
[…]
