### 2020-01-14 277e8c4
Refactors code to use E+ ExampleFiles instead of package files
