+++ tests/test_trnsys.py
@@ -14,6 +14,7 @@ from archetypal import (
     load_idf,
     settings,
     choose_window,
+    get_eplus_dirs,
 )
 
 # Function round to hundreds
@@ -51,15 +52,14 @@ class TestsConvert:
     @pytest.fixture(
         scope="class",
         params=[
-            "RefBldgWarehousePost1980_v1.4_7.2_6A_USA_MN_MINNEAPOLIS.idf",
-            "NECB 2011 - Warehouse.idf",
-            "ASHRAE90.1_Warehouse_STD2004_Rochester.idf",
-            "ASHRAE90.1_ApartmentMidRise_STD2004_Rochester.idf",
+            "RefBldgWarehouseNew2004_Chicago.idf",
+            "ASHRAE9012016_Warehouse_Denver.idf",
+            "ASHRAE9012016_ApartmentMidRise_Denver.idf",
             "5ZoneGeometryTransform.idf",
         ],
     )
     def converttest(self, config, fresh_start, request):
-        file = os.path.join("tests", "input_data", "trnsys", request.param)
+        file = get_eplus_dirs(settings.ep_version) / "ExampleFiles" / request.param
         window_file = "W74-lib.dat"
         template_dir = os.path.join("archetypal", "ressources")
         window_filepath = os.path.join(template_dir, window_file)
@@ -246,15 +246,13 @@ class TestsConvert:
 
 @pytest.fixture(
     params=[
-        "RefBldgWarehousePost1980_v1.4_7.2_6A_USA_MN_MINNEAPOLIS.idf",
-        "NECB 2011 - Warehouse.idf",
-        "ASHRAE90.1_Warehouse_STD2004_Rochester.idf",
-        "ASHRAE90.1_ApartmentMidRise_STD2004_Rochester.idf",
+        "RefBldgWarehouseNew2004_Chicago.idf",
+        "ASHRAE9012016_Warehouse_Denver.idf",
+        "ASHRAE9012016_ApartmentMidRise_Denver.idf",
     ]
 )
 def trnbuild_file(config, request):
-    file_upper_path = Path("tests") / "input_data" / "trnsys"
-    idf_file = file_upper_path / request.param
+    idf_file = get_eplus_dirs(settings.ep_version) / "ExampleFiles" / request.param
     idf_file = copy_file(idf_file, where=settings.cache_folder)
     yield idf_file
 
@@ -284,7 +282,7 @@ def test_trnbuild_from_idf(config, trnbuild_file):
     convert_idf_to_trnbuild(
         idf_file=file,
         window_lib=window_filepath,
-        template="tests/input_data/trnsys/NewFileTemplate" ".d18",
+        template="tests/input_data/trnsys/NewFileTemplate.d18",
         trnsidf_exe="docker/trnsidf/trnsidf.exe",
         **kwargs_dict
     )
@@ -297,39 +295,6 @@ def test_trnbuild_from_idf(config, trnbuild_file):
 )
 def test_trnbuild_from_idf_parallel(config, trnbuild_file):
     # All IDF files
-    idf_list = [
-        "NECB 2011 - Full Service Restaurant.idf",
-        "NECB 2011 - HighRise Apartment.idf",
-        "NECB 2011 - Hospital.idf",
-        "NECB 2011 - Large Hotel.idf",
-        "NECB 2011 - Medium Office.idf",
-        "NECB 2011 - MidRise Apartment.idf",
-        "NECB 2011 - Outpatient.idf",
[…]

[…]
