### 2020-01-15 27042f2
refactors idf_version_updater()
