### 2020-01-15 d9062af
Refactors a test
