### 2020-01-15 0cdf2ff
Refactors get_eplus_dirs()
