### 2020-01-15 1623809
New: Adds a transition method for the CLI + Test
