### 2020-01-15 4bf9cbf
Adds a new test for files coming from ClimateStudio
These files use a IdealAirLoads HVAC object which needs to be handled differently
