### 2020-01-15 5ae0056
Hides a private API key and exposes it as an environment variable for the CI
