### 2020-01-15 96d687f
Fixes some test errors
