### 2020-01-15 ca2f9f1
Fixes an issue where a shading device would be detected
