### 2020-01-15 cddea7c
Simplifies a test
