### 2020-01-15 a85acd1
Names changes in v9 affect WindowsShadingControl
