### 2020-01-17 7dba01f
Fixes an issue with construction lists
Where the first construction object would be a window element; would not work when creating an InternalMass object
