### 2020-01-17 5b09624
Improves cli command + better docs
