### 2020-01-19 cf1028f
Fixes an issue with the path being a str instead of a Path object
