### 2020-01-20 8819ed8
Fixes an issue where transition executable paths would not be correct on Windows platform
