+++ archetypal/idfclass.py
@@ -13,6 +13,7 @@ import inspect
 import logging as lg
 import multiprocessing
 import os
+import platform
 import subprocess
 import time
 from collections import defaultdict
@@ -763,7 +764,7 @@ def load_idf(
             output_folder=output_folder,
             include=include,
             epw=weather_file,
-            ep_version=ep_version,
+            ep_version=ep_version if ep_version is not None else settings.ep_version,
         )
         log("Eppy load completed in {:,.2f} seconds\n".format(time.time() - start_time))
         return idf
@@ -2320,48 +2321,49 @@ def idf_version_updater(idf_file, to_version=None, out_dir=None, simulname=None)
         vupdater_path = (
             get_eplus_dirs(settings.ep_version) / "PreProcess" / "IDFVersionUpdater"
         )
+        exe = ".exe" if platform.system() == "Windows" else ""
         trans_exec = {
-            "1-0-0": os.path.join(vupdater_path, "Transition-V1-0-0-to-V1-0-1"),
-            "1-0-1": os.path.join(vupdater_path, "Transition-V1-0-1-to-V1-0-2"),
-            "1-0-2": os.path.join(vupdater_path, "Transition-V1-0-2-to-V1-0-3"),
-            "1-0-3": os.path.join(vupdater_path, "Transition-V1-0-3-to-V1-1-0"),
-            "1-1-0": os.path.join(vupdater_path, "Transition-V1-1-0-to-V1-1-1"),
-            "1-1-1": os.path.join(vupdater_path, "Transition-V1-1-1-to-V1-2-0"),
-            "1-2-0": os.path.join(vupdater_path, "Transition-V1-2-0-to-V1-2-1"),
-            "1-2-1": os.path.join(vupdater_path, "Transition-V1-2-1-to-V1-2-2"),
-            "1-2-2": os.path.join(vupdater_path, "Transition-V1-2-2-to-V1-2-3"),
-            "1-2-3": os.path.join(vupdater_path, "Transition-V1-2-3-to-V1-3-0"),
-            "1-3-0": os.path.join(vupdater_path, "Transition-V1-3-0-to-V1-4-0"),
-            "1-4-0": os.path.join(vupdater_path, "Transition-V1-4-0-to-V2-0-0"),
-            "2-0-0": os.path.join(vupdater_path, "Transition-V2-0-0-to-V2-1-0"),
-            "2-1-0": os.path.join(vupdater_path, "Transition-V2-1-0-to-V2-2-0"),
-            "2-2-0": os.path.join(vupdater_path, "Transition-V2-2-0-to-V3-0-0"),
-            "3-0-0": os.path.join(vupdater_path, "Transition-V3-0-0-to-V3-1-0"),
-            "3-1-0": os.path.join(vupdater_path, "Transition-V3-1-0-to-V4-0-0"),
-            "4-0-0": os.path.join(vupdater_path, "Transition-V4-0-0-to-V5-0-0"),
-            "5-0-0": os.path.join(vupdater_path, "Transition-V5-0-0-to-V6-0-0"),
-            "6-0-0": os.path.join(vupdater_path, "Transition-V6-0-0-to-V7-0-0"),
-            "7-0-0": os.path.join(vupdater_path, "Transition-V7-0-0-to-V7-1-0"),
-            "7-1-0": os.path.join(vupdater_path, "Transition-V7-1-0-to-V7-2-0"),
-            "7-2-0": os.path.join(vupdater_path, "Transition-V7-2-0-to-V8-0-0"),
[…]
