### 2020-01-20 590194a
Adds Windows build script
