+++ appveyor.yml
@@ -1,0 +1,48 @@
+version: 1.0.{build}
+environment:
+  matrix:
+  - PYTHON: C:\\Python37
+    PYTHONVERSION: 3.7
+    ARCHITECTURE: i386
+    MINICONDA: C:\\Miniconda37
+    ENERGYPLUS_VERSION: 9.2.0
+    ENERGYPLUS_SHA: 921312fa1d
+  - PYTHON: C:\\Python37-x64
+    PYTHONVERSION: 3.7
+    ARCHITECTURE: x86_64
+    MINICONDA: C:\\Miniconda37-x64
+    ENERGYPLUS_VERSION: 9.2.0
+    ENERGYPLUS_SHA: 921312fa1d
+build: off
+install:
+  # set up Miniconda test environment
+  - "set PATH=%MINICONDA%;%MINICONDA%\\Scripts;%PATH%"
+  - conda config --set always_yes yes --set changeps1 no
+  - conda info -a
+  - conda update -q conda
+  - activate
+  - "conda create -n test-env python=%PYTHONVERSION% pytest"
+  - "activate test-env"
+  
+  - ECHO "Filesystem root:"
+  - ps: "ls \"C:/\""
+
+  - ECHO "Installed SDKs:"
+    - ps: "ls \"C:/Program Files/Microsoft SDKs/Windows\""
+    
+  # install EnergyPlus
+  - ECHO "Installing EnergyPlus"
+  - ps: $env:ENERGYPLUS_INSTALL_VERSION = $env:ENERGYPLUS_VERSION -replace '\.','-'
+  - ps: $env:ENERGYPLUS_DOWNLOAD_BASE_URL = "https://github.com/NREL/EnergyPlus/releases/download/v$env:ENERGYPLUS_VERSION"
+  - ps: $env:ENERGYPLUS_DOWNLOAD_FILENAME = "EnergyPlus-$env:ENERGYPLUS_VERSION-$env:ENERGYPLUS_SHA-Windows-$env:ARCHITECTURE.exe"
+  - ps: $env:ENERGYPLUS_DOWNLOAD_URL = "$env:ENERGYPLUS_DOWNLOAD_BASE_URL/$env:ENERGYPLUS_DOWNLOAD_FILENAME"
+
+  - ps: "curl $env:ENERGYPLUS_DOWNLOAD_URL -OutFile $env:TMP\\$env:ENERGYPLUS_DOWNLOAD_FILENAME"
+  - ps: "& $env:TMP\\$env:ENERGYPLUS_DOWNLOAD_FILENAME /S | Out-Null"
+
+  - ps: echo $env:PYTHON
+
+  # install archetypal
+  - "pip install .[dev]"
+test_script:
+- cmd: pytest --cov=archetypal --verbose tests/
