+++ appveyor.yml
@@ -6,14 +6,14 @@ environment:
     ARCHITECTURE: i386
     MINICONDA: C:\\Miniconda37
     ENERGYPLUS_VERSION: 9.2.0
-    ENERGYPLUS_INSTALL_VERSION=9-2-0
+    ENERGYPLUS_INSTALL_VERSION: 9-2-0
     ENERGYPLUS_SHA: 921312fa1d
   - PYTHON: C:\\Python37-x64
     PYTHONVERSION: 3.7
     ARCHITECTURE: x86_64
     MINICONDA: C:\\Miniconda37-x64
     ENERGYPLUS_VERSION: 9.2.0
-    ENERGYPLUS_INSTALL_VERSION=9-2-0
+    ENERGYPLUS_INSTALL_VERSION: 9-2-0
     ENERGYPLUS_SHA: 921312fa1d
 build: off
 install:
