### 2020-01-20 88c8b2a
Removes duplicated environment variables
