+++ appveyor.yml
@@ -3,18 +3,22 @@ environment:
   matrix:
   - PYTHON: C:\\Python37
     PYTHONVERSION: 3.7
+    PLATFORM: Windows
     ARCHITECTURE: i386
     MINICONDA: C:\\Miniconda37
     ENERGYPLUS_VERSION: 9.2.0
     ENERGYPLUS_INSTALL_VERSION: 9-2-0
     ENERGYPLUS_SHA: 921312fa1d
+    EXT: exe
   - PYTHON: C:\\Python37-x64
     PYTHONVERSION: 3.7
+    PLATFORM: Windows
     ARCHITECTURE: x86_64
     MINICONDA: C:\\Miniconda37-x64
     ENERGYPLUS_VERSION: 9.2.0
     ENERGYPLUS_INSTALL_VERSION: 9-2-0
     ENERGYPLUS_SHA: 921312fa1d
+    EXT: exe
 build: off
 install:
   # set up Miniconda test environment
@@ -33,14 +37,8 @@ install:
   - ps: "ls \"C:/Program Files/Microsoft SDKs/Windows\""
     
   # install EnergyPlus
-  - cmd: SET ENERGYPLUS_VERSION=8.9.0
-  - cmd: SET ENERGYPLUS_SHA=40101eaafd
-  - cmd: SET ENERGYPLUS_INSTALL_VERSION=8-9-0
-  - cmd: SET EXT=exe
-  - cmd: SET PLATFORM=Windows
-  
   - cmd: SET ENERGYPLUS_DOWNLOAD_BASE_URL=https://github.com/NREL/EnergyPlus/releases/download/v%ENERGYPLUS_VERSION%
-  - cmd: SET ENERGYPLUS_DOWNLOAD_FILENAME=EnergyPlus-%ENERGYPLUS_VERSION%-%ENERGYPLUS_SHA%-%PLATFORM%-x86_64
+  - cmd: SET ENERGYPLUS_DOWNLOAD_FILENAME=EnergyPlus-%ENERGYPLUS_VERSION%-%ENERGYPLUS_SHA%-%PLATFORM%-%ARCHITECTURE%
   - cmd: SET ENERGYPLUS_DOWNLOAD_URL=%ENERGYPLUS_DOWNLOAD_BASE_URL%/%ENERGYPLUS_DOWNLOAD_FILENAME%.%EXT%
   
   - cmd: ECHO Downloading EnergyPlus at...
