### 2020-01-20 4ac7277
Revert "Not verbose E+ install"
This reverts commit 0d558acc0850d4da59bbe77db13fd31bbd6bf462.
