+++ appveyor.yml
@@ -46,7 +46,7 @@ install:
   - cmd: curl -SLO %ENERGYPLUS_DOWNLOAD_URL%
   
   - cmd: ECHO Installing EnergyPlus...
-  - cmd: "%ENERGYPLUS_DOWNLOAD_FILENAME% /S"
+  - cmd: "%ENERGYPLUS_DOWNLOAD_FILENAME% /S | Out-Null"
   - cmd: ECHO Done
   
     # Install additional material
