### 2020-01-20 22a61da
Fixes % for $
