+++ appveyor.yml
@@ -38,15 +38,15 @@ install:
     
   # install EnergyPlus
   - cmd: SET ENERGYPLUS_DOWNLOAD_BASE_URL=https://github.com/NREL/EnergyPlus/releases/download/v%ENERGYPLUS_VERSION%
-  - cmd: SET ENERGYPLUS_DOWNLOAD_FILENAME=EnergyPlus-%ENERGYPLUS_VERSION%-%ENERGYPLUS_SHA%-%PLATFORM%-%ARCHITECTURE%
-  - cmd: SET ENERGYPLUS_DOWNLOAD_URL=%ENERGYPLUS_DOWNLOAD_BASE_URL%/%ENERGYPLUS_DOWNLOAD_FILENAME%.%EXT%
+  - cmd: SET ENERGYPLUS_DOWNLOAD_FILENAME=EnergyPlus-%ENERGYPLUS_VERSION%-%ENERGYPLUS_SHA%-%PLATFORM%-%ARCHITECTURE%.%EXT%
+  - cmd: SET ENERGYPLUS_DOWNLOAD_URL=%ENERGYPLUS_DOWNLOAD_BASE_URL%/%ENERGYPLUS_DOWNLOAD_FILENAME%
   
   - cmd: ECHO Downloading EnergyPlus at...
   - cmd: ECHO %ENERGYPLUS_DOWNLOAD_URL%
   - cmd: curl -SLO %ENERGYPLUS_DOWNLOAD_URL%
   
   - cmd: ECHO Installing EnergyPlus...
-  - ps: "& $env:ENERGYPLUS_DOWNLOAD_FILENAME.$env:EXT /S | Out-Null"
+  - ps: "& $env:ENERGYPLUS_DOWNLOAD_FILENAME /S | Out-Null"
   - cmd: ECHO Done
   
     # Install additional material
