### 2020-01-21 a9545eb
Sorting schedule names before writing it in output file
