### 2020-01-24 acc6e19
Fixes an issue where an install_script is necessary for a silent install
