### 2020-01-24 ecd5293
quotes issue
