+++ appveyor.yml
@@ -42,12 +42,12 @@ install:
   
   - cmd: ECHO Downloading EnergyPlus at...
   - cmd: ECHO %ENERGYPLUS_DOWNLOAD_URL%
-  - cmd: curl -SLO %ENERGYPLUS_DOWNLOAD_URL% -O %TMP%\%ENERGYPLUS_DOWNLOAD_FILENAME%
+  - cmd: curl -SLO %ENERGYPLUS_DOWNLOAD_URL% -O %TMP%\\%ENERGYPLUS_DOWNLOAD_FILENAME%
   
   # getting install script for silent install
   - cmd: SET INSTALLSCRIPT_URL=https://raw.githubusercontent.com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs
-  - cmd: "curl -SLO %INSTALLSCRIPT_URL% -O %TMP%\install_script.qs"
-  - cmd: "%TMP%\%ENERGYPLUS_DOWNLOAD_FILENAME% --verbose --script %TMP%\install_script.qs"
+  - cmd: "curl -SLO %INSTALLSCRIPT_URL% -O %TMP%\\install_script.qs"
+  - cmd: "%TMP%\\%ENERGYPLUS_DOWNLOAD_FILENAME% --verbose --script %TMP%\\install_script.qs"
   
     # Install additional material
   - ECHO Installing additional files...
@@ -62,8 +62,8 @@ install:
   - SET DEST=C:\EnergyPlusV%ENERGYPLUS_INSTALL_VERSION%\PreProcess\IDFVersionUpdater
   - ECHO Extracting and Copying files to...
   - ECHO %DEST%
-  - tar xvf %TMP%\%ATTCHNUM% -C %DEST%
-  - DEL /s %TMP%\%ATTCHNUM%
+  - tar xvf %TMP%\\%ATTCHNUM% -C %DEST%
+  - DEL /s %TMP%\\%ATTCHNUM%
   
   - ECHO Succesfully installed EnergyPlus and conversion files
 
