+++ appveyor.yml
@@ -57,13 +57,13 @@ install:
   
   - ECHO Installing older conversion files from...
   - ECHO %EXTRAS_DOWNLOAD_URL%
-  - curl -SLO %EXTRAS_DOWNLOAD_URL% -O %TMP%\%ATTCHNUM%
+  - curl -SLO %EXTRAS_DOWNLOAD_URL%
   
   - SET DEST=C:\EnergyPlusV%ENERGYPLUS_INSTALL_VERSION%\PreProcess\IDFVersionUpdater
   - ECHO Extracting and Copying files to...
   - ECHO %DEST%
-  - tar xvf %TMP%\\%ATTCHNUM% -C %DEST%
-  - DEL /s %TMP%\\%ATTCHNUM%
+  - tar xvf %ATTCHNUM% -C %DEST%
+  - DEL /s %ATTCHNUM%
   
   - ECHO Succesfully installed EnergyPlus and conversion files
 
