### 2020-01-27 680b5bc
Adds the extension to attachment
