### 2020-01-27 2b589dc
Adds -Force option to Expand-Archive
