### 2020-01-27 e4d8510
Installs Shapely using wheels
