+++ appveyor.yml
@@ -73,6 +73,10 @@ install:
   - ECHO Succesfully installed EnergyPlus and conversion files
 
   - ECHO %PYTHON%
+  
+  - curl -SLO https://download.lfd.uci.edu/pythonlibs/q4hpdf1k/Shapely-1.6.4.post2-cp37-cp37m-win_amd64.whl
+  - pip wheel Shapely-1.6.4.post2-cp37-cp37m-win_amd64.whl
+  - DEL /s Shapely-1.6.4.post2-cp37-cp37m-win_amd64.whl
 
   # install archetypal
   - "pip install .[dev]"
