### 2020-01-27 5ea0dc9
Install shapely using conda instead
