### 2020-01-27 30ae6b0
Generalize xfails on CI environments
