### 2020-01-27 a62df0d
No need to remove the tmp folder in _run_exec() since it is handled in the parent method
