### 2020-01-28 43e657a
Fixes an issue where the platform number would not cast to an int
