### 2020-01-28 540e4c4
Adds Windows CI (#59)
* Adds Windows build script

* Typo

* Trying using CMD

* Trying using CMD (fixed typo)

* Maybe

* Removes duplicated environment variables

* Not verbose E+ install

* Test

* Revert "Test"

This reverts commit 56896cd92ffe23c4ef171b9ded8ee22969451a55.

* Revert "Not verbose E+ install"

This reverts commit 0d558acc0850d4da59bbe77db13fd31bbd6bf462.

* Test2

* Using PS instead of CMD

* Fixes '/' operator

* Adds extension to executable path

* Fixes % for $

* Trying to fix path error

* back to cmd

* Back to original command

* Build at top

* Trying something

* sm
[…]
