+++ tests/test_trnsys.py
@@ -258,8 +258,8 @@ def trnbuild_file(config, request):
 
 
 @pytest.mark.xfail(
-    "TRAVIS" in os.environ and os.environ["TRAVIS"] == "true",
-    reason="Skipping this test on Travis CI.",
+    os.environ["CI"].lower() == "true",
+    reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_from_idf(config, trnbuild_file):
     # List files here
@@ -290,8 +290,8 @@ def test_trnbuild_from_idf(config, trnbuild_file):
 
 @pytest.mark.win32
 @pytest.mark.xfail(
-    "TRAVIS" in os.environ and os.environ["TRAVIS"] == "true",
-    reason="Skipping this test on Travis CI.",
+    os.environ["CI"].lower() == "true",
+    reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_from_idf_parallel(config, trnbuild_file):
     # All IDF files
@@ -316,8 +316,8 @@ def test_trnbuild_from_idf_parallel(config, trnbuild_file):
 @pytest.mark.darwin
 @pytest.mark.linux
 @pytest.mark.xfail(
-    "TRAVIS" in os.environ and os.environ["TRAVIS"] == "true",
-    reason="Skipping this test on Travis CI.",
+    os.environ["CI"].lower() == "true",
+    reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_from_idf_parallel_darwin_or_linux(config):
     # All IDF files
@@ -347,8 +347,8 @@ def test_trnbuild_from_idf_parallel_darwin_or_linux(config):
 
 @pytest.mark.win32
 @pytest.mark.xfail(
-    "TRAVIS" in os.environ and os.environ["TRAVIS"] == "true",
-    reason="Skipping this test on Travis CI.",
+    os.environ["CI"].lower() == "true",
+    reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_idf_win32(config):
     idf_file = "tests/input_data/trnsys/Building.idf"
@@ -361,8 +361,8 @@ def test_trnbuild_idf_win32(config):
 @pytest.mark.darwin
 @pytest.mark.linux
 @pytest.mark.xfail(
-    "TRAVIS" in os.environ and os.environ["TRAVIS"] == "true",
-    reason="Skipping this test on Travis CI.",
+    os.environ["CI"].lower() == "true",
+    reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_idf_darwin_or_linux(config):
     idf_file = "tests/input_data/trnsys/Building.idf"
