+++ appveyor.yml
@@ -1,0 +1,83 @@
+build: false
+environment:
+  matrix:
+  - PYTHON: "C:\\Python37"
+    PYTHONVERSION: 3.7
+    PLATFORM: "Windows"
+    ARCHITECTURE: "i386"
+    MINICONDA: "C:\\Miniconda37"
+    ENERGYPLUS_VERSION: "9.2.0"
+    ENERGYPLUS_INSTALL_VERSION: "9-2-0"
+    ENERGYPLUS_SHA: "921312fa1d"
+    EXT: "exe"
+  - PYTHON: "C:\\Python37-x64"
+    PYTHONVERSION: 3.7
+    PLATFORM: "Windows"
+    ARCHITECTURE: "x86_64"
+    MINICONDA: "C:\\Miniconda37-x64"
+    ENERGYPLUS_VERSION: "9.2.0"
+    ENERGYPLUS_INSTALL_VERSION: "9-2-0"
+    ENERGYPLUS_SHA: "921312fa1d"
+    EXT: "exe"
+install:
+  # set up Miniconda test environment
+  - "set PATH=%MINICONDA%;%MINICONDA%\\Scripts;%PATH%"
+  - conda config --set always_yes yes --set changeps1 no
+  - conda info -a
+  - conda update -q conda
+  - activate
+  - "conda create -n test-env python=%PYTHONVERSION% pytest"
+  - "activate test-env"
+  
+  - ECHO "Filesystem root:"
+  - ps: "ls \"C:/\""
+
+  - ECHO "Installed SDKs:"
+  - ps: "ls \"C:/Program Files/Microsoft SDKs/Windows\""
+    
+  # install EnergyPlus
+  - cmd: SET ENERGYPLUS_DOWNLOAD_BASE_URL=https://github.com/NREL/EnergyPlus/releases/download/v%ENERGYPLUS_VERSION%
+  - cmd: SET ENERGYPLUS_DOWNLOAD_FILENAME=EnergyPlus-%ENERGYPLUS_VERSION%-%ENERGYPLUS_SHA%-%PLATFORM%-%ARCHITECTURE%.%EXT%
+  - cmd: SET ENERGYPLUS_DOWNLOAD_URL=%ENERGYPLUS_DOWNLOAD_BASE_URL%/%ENERGYPLUS_DOWNLOAD_FILENAME%
+  
+  - cmd: ECHO Downloading EnergyPlus at...
+  - cmd: ECHO %ENERGYPLUS_DOWNLOAD_URL%
+  - cmd: curl -SLO %ENERGYPLUS_DOWNLOAD_URL%
+  
+  # getting install script for silent install
+  - cmd: SET INSTALLSCRIPT_URL=https://raw.githubusercontent.com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs
+  - cmd: "curl -SLO %INSTALLSCRIPT_URL%"
+  - cmd: "%ENERGYPLUS_DOWNLOAD_FILENAME% --verbose --script install_script.qs"
+  
+    # Install additional material
+  - ECHO Installing additional files...
+  - SET ATTCHBASE=86
+  - SET ATTCHNUM=8231
+  - SET EXTRAS_DOWNLOAD_URL="http://energyplus.helpserve.com/Knowledgebase/Article/GetAttachment/%ATTCHBASE%/%ATTCHNUM%"
+  
+  - ECHO Installing older conversion files from...
+  - ECHO %EXTRAS_DOWNLOAD_URL%
+  - curl -SL %EXTRAS_DOWNLOAD_URL% -o %ATTCHNUM%.zip
+  
+  - ECHO Listing all files in current directory
+  - ps: ls
+  
[…]
