### 2020-01-28 f80047d
Fixes doc so that it runs in console
