### 2020-01-28 bb4b3fb
Code cleanup
