### 2020-01-28 0b20b7b
Better UMI converter docs
