### 2020-01-28 d135025
Better default output file
