### 2020-01-28 bbacb7c
Fixes curl path
