### 2020-01-28 fab7a14
Handles more gracefully transitions when files are not available
