### 2020-01-28 29f9d53
Removes --verbose option for pytest
