+++ .travis.yml
@@ -92,7 +92,7 @@ install:
   - pip install --upgrade pip
   - pip install .[dev]
 script:
-  - py.test --cov=archetypal --verbose tests/
+  - py.test --cov=archetypal tests/
 after_success:
   - coverage report -m
   - coveralls
