### 2020-01-28 ffee68a
Fixes extraction path
