### 2020-01-28 60a7eca
Fixes path in method reference in docs
