### 2020-01-29 5b96874
Adds DataPortal module documentation
