### 2020-01-29 471e61e
Removes the first.rst file that is not necessary
