+++ docs/first.rst
@@ -1,6 +1,0 @@
-Getting Started
-===============
-
-.. todo::
-
-    This section is not yet available
