### 2020-01-29 9d12737
Refactors documentation in a better structure
