+++ docs/tutorials.rst
@@ -1,8 +1,0 @@
-Tutorials
-=========
-
-.. toctree::
-    :maxdepth: 1
-    :glob:
-
-    tutorials/*
