+++ docs/converting.rst
@@ -1,0 +1,6 @@
+Converting IDF models
+=====================
+
+.. include:: converter_bui.rst
+
+.. include:: converter_umi.rst
