+++ docs/converting.rst
@@ -1,6 +1,9 @@
 Converting IDF models
 =====================
 
+EnergyPlus models can be converted to two different other formats using `archetypal`. TRNSYS users can convert IDF
+files to TRNBuild/Type56 files while UMI users can convert IDF files to UMI Template Files.
+
 .. include:: converter_bui.rst
 
 .. include:: converter_umi.rst
