### 2020-01-30 073d595
Updates dependencies
This relates to issue #42
