+++ requirements.txt
@@ -10,7 +10,7 @@ scikit-learn
 scipy
 pandas>=0.24
 pyomo
-numpy>=1.13
+numpy>=1.17
 tqdm
 tabulate
 path.py
@@ -18,4 +18,3 @@ tsam
 pint
 click
 outdated
-path.py
