### 2020-01-30 7349954
Adds pip to conda requirements
