### 2020-01-30 9c09456
Adds conda installation procedure
This fixes issue #42
