+++ docs/install.rst
@@ -10,8 +10,8 @@ Prior to installing this package, you must have the 8.9.0 version (download_ her
 would be in `/Applications/EnergyPlus-8-9-0`. For a step-by-step installation of EnergyPlus, see `Installation from
 scratch (Windows)`_
 
-Installation with ``pip``
--------------------------
+Installing using ``pip``
+------------------------
 
 If you have Python 3 already installed on your machine and don't bother to create a virtual environment (which is
 highly recommended), then simply install using the following command in the terminal:
@@ -19,6 +19,11 @@ highly recommended), then simply install using the following command in the term
 .. code-block:: shell
 
     pip install archetypal
+
+.. hint::
+
+    If you encounter an issue during the installation of archetypal using ``pip``, you can try
+    out `Installing using conda (Anaconda)`_ instead.
 
 
 Installation within a Virtual Environment
@@ -77,21 +82,25 @@ project, *archetypal* in this case.
    and select the appropriate virtual env created earlier (*archetypal* in this case).
 
 
-Installation with Anaconda / conda
-----------------------------------
+Installing using ``conda`` (Anaconda)
+-------------------------------------
 
-.. warning::
+.. hint::
 
-    This install method  is not yet available. Revert to the method detailed above.
+    If you encounter package dependency errors while installing `archetypal` using pip, you can use conda instead.
 
-
-Installing with conda is similar:
+Installing with conda is similar to pip. The following workflow creates a new virtual environment (named archetypal)
+which contains the required dependencies. It then installs the package using pip. You will need to download the
+`environment.yml`_ file from the github repository. For the following code to work, first change the working
+directory to the location of the downloaded `environment.yml` file. Here we use the `conda env update` method which
+will work well to create a new environment using a specific dependency file in one line of code:
 
 .. code-block:: shell
 
    conda update -n base conda
-   conda create -n archetypal python=3 archetypal
-   source activate archetypal
+   conda env update -n archetypal -f environment.yml
+   conda activate archetypal
+   pip install archetypal
 
 
 Installation from scratch (Windows)
@@ -151,7 +160,7 @@ whole code block below in Command Prompt and Hit :guilabel:`&Enter:⏎`.
     cd aef233396167e0f961df3d62a193573e
     install_eplus_script.cmd
 
-To install *archetypal*, follow the steps detailes above in `Installation with pip`_
+To install *archetypal*, follow the steps detailed above in `Installing using pip`_
 
 
 .. _start a jupyter notebook: https://jupyter.readthedocs.io/en/latest/running.html#starting-the-notebook-server
@@ -163,3 +172,4 @@ To install *archetypal*, follow the steps detailes above in `Installation with p
 .. _download: https://github.com/NREL/EnergyPlus/releases/tag/v8.9.0
 .. _supplementary conversion programs: http://energyplus.helpserve.com/Knowledgebase/List/Index/46/converting-older-version-files
 .. _script: https://gist.github.com/samuelduchesne/aef233396167e0f961df3d62a193573e
+.. _environment.yml: https://github.com/samuelduchesne/archetypal/blob/master/environment.yml
