### 2020-01-30 e2e1cc0
Removes unnecessary with statement
This relates to issue #46
