### 2020-01-30 samuelduchesne/archetypal#46 @samuelduchesne
@louisleroy5, I am getting an error when I run `archetypal convert "tests/input_data/trnsys/test_multizone_EP.idf" "tests/.temp"` in the command line inside Pycharm. Can you check this out?

The error is:
```
Traceback (most recent call last):
  File "/Users/samuelduchesne/anaconda3/envs/archetypal/bin/archetypal", line 11, in <module>
    load_entry_point('archetypal', 'console_scripts', 'archetypal')()
  File "/Users/samuelduchesne/anaconda3/envs/archetypal/lib/python3.7/site-packages/click/core.py", line 764, in __call__
    return self.main(*args, **kwargs)
  File "/Users/samuelduchesne/anaconda3/envs/archetypal/lib/python3.7/site-packages/click/core.py", line 717, in main
    rv = self.invoke(ctx)
  File "/Users/samuelduchesne/anaconda3/envs/archetypal/lib/python3.7/site-packages/click/core.py", line 1137, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/Users/samuelduchesne/anaconda3/envs/archetypal/lib/python3.7/site-packages/click/core.py", line 956, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/Users/samuelduchesne/anaconda3/envs/archetypal/lib/python3.7/site-packages/click/core.py", line 555, in invoke
    return callback(*args, **kwargs)
  File "/Users/samuelduchesne/Dropbox/Polytechnique/Doc/software/archetypal/archetypal/cli.py", line 294, in convert
    capacitance=capacitance
  File "/Users/samuelduchesne/Dropbox/Polytechnique/Doc/software/archetypal/archetypal/trnsys.py", line 105, in convert_idf_to_trnbuild
    idf_file, window_lib, output_folder, trnsidf_exe, template
  File "/Users/samuelduchesne/Dropbox/Polytechnique/Doc/software/archetypal/archetypal/trnsys.py", line 560, in _assert_files
    raise IOError("idf_file file not found")
OSError: idf_file file not found

```

I am running this on `develop`
