### 2020-01-30 samuelduchesne/archetypal#44 @samuelduchesne
This is fixed by #60 and #62
