### 2020-01-30 2191ec0
Adds a more prominent message about the e+ version requirement
