### 2020-01-30 samuelduchesne/archetypal#43 @samuelduchesne
Since #58, the package requirement is EnergyPlus 9.2.0 and above. This simplifies a lot of things. The documentation reflects this in 2191ec00175310b9f35e2f9e53e55a902c2a0892
