### 2020-01-30 019b5a3
Fix: uses "verbose" instead of "--log-console"
