+++ tests/test_cli.py
@@ -24,7 +24,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -44,7 +44,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -67,7 +67,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -89,7 +89,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -111,7 +111,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -132,7 +132,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -159,7 +159,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -180,7 +180,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
@@ -207,7 +207,7 @@ class TestCli:
                 "tests/.temp/images",
                 "--logs-folder",
                 "tests/.temp/logs",
-                "--log-console",
+                "--verbose",
                 "convert",
                 "--ep-version",
                 "9-1-0",
