### 2020-01-30 da3d9b5
Fix: new way to get os.environ "CI"
