### 2020-01-30 ca29ee2
Fix: uses settings.ep_version in tests instead of string
