### 2020-01-30 589fc42
Fix: uses settings.ep_version in tests instead of string
