+++ tests/test_trnsys.py
@@ -518,7 +518,7 @@ def test_trnbuild_from_idf(config, trnbuild_file):
     # prepare args (key=value). Key is a unique id for the runs (here the
     # file basename is used). Value is a dict of the function arguments
     kwargs_dict = {
-        "ep_version": "8-9-0",
+        "ep_version": settings.ep_version,
         "u_value": 2.5,
         "shgc": 0.6,
         "t_vis": 0.78,
