### 2020-01-30 dadca05
Fix: getting CI env from os.environ in more robust way
