+++ tests/test_trnsys.py
@@ -502,7 +502,7 @@ def trnbuild_file(config, request):
 
 
 @pytest.mark.xfail(
-    os.environ.get("CI", True),
+    os.environ.get("CI", False).lower() == "true",
     reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_from_idf(config, trnbuild_file):
@@ -541,7 +541,7 @@ def test_trnbuild_from_idf(config, trnbuild_file):
 
 @pytest.mark.win32
 @pytest.mark.xfail(
-    os.environ.get("CI", True),
+    os.environ.get("CI", False).lower() == "true",
     reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_from_idf_parallel(config, trnbuild_file):
@@ -571,7 +571,7 @@ def test_trnbuild_from_idf_parallel(config, trnbuild_file):
 @pytest.mark.darwin
 @pytest.mark.linux
 @pytest.mark.xfail(
-    os.environ.get("CI", True),
+    os.environ.get("CI", False).lower() == "true",
     reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_from_idf_parallel_darwin_or_linux(config):
@@ -608,7 +608,7 @@ def test_trnbuild_from_idf_parallel_darwin_or_linux(config):
 
 @pytest.mark.win32
 @pytest.mark.xfail(
-    os.environ.get("CI", True),
+    os.environ.get("CI", False).lower() == "true",
     reason="Skipping this test on CI environment.",
 )
 def test_trnbuild_idf_win32(config):
@@ -639,7 +639,7 @@ def safe_int_cast(val, default=0):
 @pytest.mark.darwin
 @pytest.mark.linux
 @pytest.mark.skipif(
-    os.environ.get("CI", True),
+    os.environ.get("CI", False).lower() == "true",
     reason="Skipping this test on CI environment.",
 )
 @pytest.mark.xfail(
