### 2020-01-30 236fb8a
Actually fixes issue #46
The with statement is not necessary
