### 2020-01-30 1c9b9e0
Reordered and streamlined cli command names
Also reflected changes in the documentation
