+++ tests/test_cli.py
@@ -26,11 +26,11 @@ class TestCli:
                 "tests/.temp/logs",
                 "--verbose",
                 "convert",
-                "--ep-version",
-                "9-1-0",
+                "--ep_version",
+                "9-2-0",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
-                "--trnsidf-exe",
+                "--trnsidf_exe",
                 "docker/trnsidf/trnsidf.exe",
             ],
             [
@@ -46,14 +46,14 @@ class TestCli:
                 "tests/.temp/logs",
                 "--verbose",
                 "convert",
-                "--ep-version",
-                "9-1-0",
+                "--ep_version",
+                "9-2-0",
                 "-i",
                 "-t",
                 "-d",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
-                "--trnsidf-exe",
+                "--trnsidf_exe",
                 "docker/trnsidf/trnsidf.exe",
             ],
             [
@@ -69,13 +69,13 @@ class TestCli:
                 "tests/.temp/logs",
                 "--verbose",
                 "convert",
-                "--ep-version",
-                "9-1-0",
+                "--ep_version",
+                "9-2-0",
                 "--window-lib",
                 "archetypal/ressources/W74-lib.dat",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
-                "--trnsidf-exe",
+                "--trnsidf_exe",
                 "docker/trnsidf/trnsidf.exe",
             ],
             [
@@ -91,13 +91,13 @@ class TestCli:
                 "tests/.temp/logs",
                 "--verbose",
                 "convert",
-                "--ep-version",
-                "9-1-0",
+                "--ep_version",
+                "9-2-0",
                 "--template",
                 "archetypal/ressources/NewFileTemplate.d18",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
-                "--trnsidf-exe",
+                "--trnsidf_exe",
                 "docker/trnsidf/trnsidf.exe",
             ],
             [
@@ -113,12 +113,12 @@ class TestCli:
                 "tests/.temp/logs",
                 "--verbose",
                 "convert",
-                "--ep-version",
-                "9-1-0",
+                "--ep_version",
+                "9-2-0",
                 "--log-clear-names",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
-                "--trnsidf-exe",
+                "--trnsidf_exe",
                 "docker/trnsidf/trnsidf.exe",
             ],
             [
@@ -134,8 +134,8 @@ class TestCli:
                 "tests/.temp/logs",
                 "--verbose",
                 "convert",
-                "--ep-version",
-                "9-1-0",
+                "--ep_version",
+                "9-2-0",
                 "--window",
                 1.5,
                 0.6,
@@ -145,7 +145,7 @@ class TestCli:
                 5.6,
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
-                "--trnsidf-exe",
+                "--trnsidf_exe",
                 "docker/trnsidf/trnsidf.exe",
[…]
