### 2020-01-31 6773de0
Uses pandas.data_range instead of DateTimeIndex
