### 2020-01-31 3f6f286
Environ CI test must return a string and not a bool
