### 2020-01-31 03c4ba5
Skips convert test on macOS catalina and above because can't use wine
