### 2020-01-31 cf94f2a
Refactors conda env creation in travis CI
