### 2020-01-31 9856e75
should fix conda init issue on travis-ci
