### 2020-01-31 ab4fea8
Simplifies travis-ci file
