+++ archetypal/template/schedule.py
@@ -143,10 +143,7 @@ class UmiSchedule(Schedule, UmiBase, metaclass=Unique):
             )
         else:
             new_values = np.average(
-                [
-                    self.all_values * quantity[0],
-                    other.all_values * quantity[1],
-                ],
+                [self.all_values * quantity[0], other.all_values * quantity[1]],
                 axis=0,
                 weights=weights,
             )
@@ -219,16 +216,25 @@ class UmiSchedule(Schedule, UmiBase, metaclass=Unique):
 
 
 class YearScheduleParts:
+    """Helper Class for YearSchedules that are defined using FromDay FromMonth
+    ToDay ToMonth attributes.
+    """
+
     def __init__(
         self, FromDay=None, FromMonth=None, ToDay=None, ToMonth=None, Schedule=None
     ):
         """
         Args:
-            FromDay:
-            FromMonth:
-            ToDay:
-            ToMonth:
-            Schedule:
+            FromDay (int): This numeric field is the starting day for the
+                schedule time period.
+            FromMonth (int): This numeric field is the starting month for the
+                schedule time period.
+            ToDay (int): This numeric field is the ending day for the schedule
+                time period.
+            ToMonth (int): This numeric field is the ending month for the
+                schedule time period.
+            Schedule (UmiSchedule): The associated UmiSchedule related to this
+                object.
         """
         self.FromDay = FromDay
         self.FromMonth = FromMonth
@@ -264,24 +270,20 @@ class YearScheduleParts:
 
 
 class DaySchedule(UmiSchedule):
-    """A DaySchedule is a superclass of UmiSchedule and handles the conversion
-    to and from the json UmiTemplate.
-    """
+    """Superclass of UmiSchedule that handles daily schedules."""
 
     def __init__(self, **kwargs):
-        """Initialize a DaySchedule object with parameters: :param ** kwargs:
-        keywords passed to the :class:`UmiSchedule`:param constructor. See
-        :class:`UmiSchedule` for more info.:
+        """Initialize a DaySchedule object with parameters:
 
         Args:
             **kwargs: Keywords passed to the :class:`UmiSchedule` constructor.
-                See :class:`UmiSchedule` for more details.
         """
         super(DaySchedule, self).__init__(**kwargs)
 
     @classmethod
     def from_epbunch(cls, epbunch, **kwargs):
-        """Create a DaySchedule from a :class:`EpBunch` object
+        """Create a DaySchedule from a :class:`~eppy.bunch_subclass.EpBunch`
+        object
 
         Args:
             epbunch (EpBunch): The EpBunch object to construct a DaySchedule
@@ -316,9 +318,10 @@ class DaySchedule(UmiSchedule):
 
     @classmethod
     def from_json(cls, Type, **kwargs):
-        """
+        """Create a DaySchedule from a Umi Template json file.
+
         Args:
-            Type:
+            Type (str): The schedule type limits name.
             **kwargs:
         """
         values = kwargs.pop("Values")
@@ -327,7 +330,13 @@ class DaySchedule(UmiSchedule):
         return sched
 
     def to_json(self):
-        """Convert class properties to dict"""
+        """Returns a dict-like representation of the schedule.
+
+        Returns:
+            dict: The dict-like representation of the schedule
+
+        """
+
         data_dict = collections.OrderedDict()
 
         data_dict["$id"] = str(self.id)
@@ -345,18 +354,18 @@ class DaySchedule(UmiSchedule):
         return np.array(self.values)
 
     def to_dict(self):
+        """returns umi template repr"""
         return {"$ref": str(self.id)}
 
 
 class WeekSchedule(UmiSchedule):
[…]
