### 2020-01-31 5707270
Should be a DateTimeIndex object instead of calling date_range method
