### 2020-02-03 a333a23
Removes some methods from docs
