### 2020-02-03 44d9e85
Adds info on caching system in docs
