### 2020-02-03 aa57bd7
Adds info on plotting schedules in docs
