+++ docs/tutorials/schedules.rst
@@ -1,10 +1,8 @@
 Schedules
 =========
 
-*archetypal* can parse schedules native to EnergyPlus Models. In EnergyPlus, there are many ways to define schedules
-but sometimes, it is necessary to have a single schedule representation. This is the case for the EnergyPlus to BUI
-converter and for the upcoming UMI converter. In both cases, the schedules are defined as a Yearly, Weekly, Daily
-schedule object. The Schedule module of *archetypal* handles this conversion.
+`archetypal` can parse EnergyPlus schedules. In EnergyPlus, there are many ways to define schedules in an IDF file. The
+Schedule module defines a class that handles parsing, plotting converting schedules.
 
 Reading Schedules
 -----------------
@@ -13,13 +11,33 @@ Reading Schedules
 
 .. code-block:: python
 
-    import archetypal as ar
-    idf = ar.load_idf(<idf-file-path>)
-    this_schedule = Schedule(Name='name', idf=idf)
+    >>> import archetypal as ar
+    >>> idf = ar.load_idf(<idf-file-path>)
+    >>> this_schedule = Schedule(Name='name', idf=idf)
 
-One can create the `year-week-day` representation for any schedule object by invoking
-the :py:meth:`archetypal.schedule.Schedule.to_year_week_day` method:
+
+Converting Schedules
+--------------------
+
+Some tools typically rely on a group of 3 schedules; defined as a Yearly, Weekly, Daily schedule object. This is the
+case for the :ref:`IDF to UMI <Converting IDF to UMI>` converter and for the :ref:`IDF to TRNSYS <Converting IDF to BUI>`
+converter. The Schedule module of *archetypal* can handle this conversion.
+
+The `year-week-day` representation for any schedule object is invoked with
+the :py:meth:`~archetypal.schedule.Schedule.to_year_week_day` method:
 
 .. code-block:: python
 
-    this_schedule.to_year_week_day()
+    >>> this_schedule.to_year_week_day()
+
+Plotting Schedules
+------------------
+
+Schedules can be parsed as :class:`pandas.Series` objects (call the `series` property on a Schedule object) which then
+exposes useful methods from the pandas package. For convenience, a wrapper for the plotting method is built-in the
+Schedule class. To plot the full annual schedule (or a specific range), simply call the :meth:`archetypal.schedule.Schedule.plot`
+method. For example,
+
+.. code-block:: python
+
+    >>> this_schedule.plot(slice=("2018/01/02", "2018/01/03"), drawstyle="steps-post")
