+++ docs/index.rst
@@ -42,9 +42,9 @@ independently.
    :caption: User Guide
 
    Converting IDF models <converting.rst>
-   Managing Schedules <tutorials/schedules.rst>
    Reading and Running IDF files <reading_idf.rst>
    Parallel Processing <tutorials/parallel_process.rst>
+   Managing Schedules <tutorials/schedules.rst>
    Troubleshooting <troubleshooting.rst>
 
 .. toctree::
