### 2020-02-03 9164b14
Refactors some dataportal methods to use parameter arguments instead of passing a dict
Also has a lot more docstring
