### 2020-02-03 c2823b3
Addresses comments under "paper summary comments" and "Misc comments on the paper" in issue #41
