### 2020-02-03 bd2f07d
refactors UmiTemplate.from_json to UmiTemplate.read_file
