### 2020-02-03 09886ce
Removes deprecated method
