### 2020-02-03 e734457
Exposes UmiTemplate class instead of methods
