### 2020-02-03 22a70fd
New: documentation for reading and editing UMI Template Files
