+++ docs/index.rst
@@ -45,6 +45,7 @@ independently.
    Reading and Running IDF files <reading_idf.rst>
    Parallel Processing <tutorials/parallel_process.rst>
    Managing Schedules <tutorials/schedules.rst>
+   Reading and Editing UMI Template Files <tutorials/edit_umitemplate.rst>
    Troubleshooting <troubleshooting.rst>
 
 .. toctree::
