### 2020-02-05 8efafbd
Refactors the UmiTemplate class to fit the latest logic
