+++ archetypal/umi_template.py
@@ -4,6 +4,7 @@ import os
 from collections import OrderedDict
 
 import numpy as np
+from path import Path
 
 from archetypal import (
     load_idf,
@@ -11,8 +12,6 @@ from archetypal import (
     GasMaterial,
     GlazingMaterial,
     OpaqueMaterial,
-    parallel_process,
-    run_eplus,
     OpaqueConstruction,
     WindowConstruction,
     StructureDefinition,
@@ -35,10 +34,9 @@ from archetypal import (
 
 
 class UmiTemplate:
+    """Main class supporting the definition of a multiple building templates and
+    corresponding template objects.
     """
-
-    """
-
     def __init__(
         self,
         name="unnamed",
@@ -60,27 +58,38 @@ class UmiTemplate:
         ZoneLoads=None,
         Zones=None,
     ):
-        """
+        """Initialize a new UmiTemplate with empty attributes.
 
         Args:
-            name (str): The name of the template
-            Zones (list of Zone):
-            ZoneLoads (list of ZoneLoad):
-            ZoneConstructionSets (list of ZoneConstructionSet):
-            ZoneConditionings (list of ZoneConditioning):
-            WindowSettings (list of WindowSetting):
-            VentilationSettings (list of VentilationSetting):
-            DomesticHotWaterSettings (list of DomesticHotWaterSetting):
-            YearSchedules (list of YearSchedule):
-            WeekSchedules (list of WeekSchedule):
-            DaySchedules (list of DaySchedule):
-            StructureDefinitions (list of StructureDefinition):
-            WindowConstructions (list of WindowConstruction):
-            OpaqueMaterials (list of OpaqueMaterial):
-            OpaqueConstructions (list of OpaqueConstruction):
-            GlazingMaterials (list of GlazingMaterial):
-            GasMaterials (list of GasMaterial):
-            BuildingTemplates (list of BuildingTemplate):
+            name (str): The name of the UMI Template.
+            BuildingTemplates (list of BuildingTemplate): list of
+                BuildingTemplate objects.
+            GasMaterials (list of GasMaterial): list of GasMaterial objects.
+            GlazingMaterials (list of GlazingMaterial): list of GlazingMaterial
+                objects.
+            OpaqueConstructions (list of OpaqueConstruction): list of
+                OpaqueConstruction objects.
+            OpaqueMaterials (list of OpaqueMaterial): list of OpaqueMaterial
+                objects.
+            WindowConstructions (list of WindowConstruction): list of
+                WindowConstruction objects.
+            StructureDefinitions (list of StructureDefinition): list of
+                StructureDefinition objects.
+            DaySchedules (list of DaySchedule): list of DaySchedule objects.
+            WeekSchedules (list of WeekSchedule): list of WeekSchedule objects.
+            YearSchedules (list of YearSchedule): list of YearSchedule objects.
+            DomesticHotWaterSettings (list of DomesticHotWaterSetting): list of
+                DomesticHotWaterSetting objects.
+            VentilationSettings (list of VentilationSetting): list of
+                VentilationSetting objects.
+            WindowSettings (list of WindowSetting): list of WindowSetting
+                objects.
+            ZoneConditionings (list of ZoneConditioning): list of
+                ZoneConditioning objects.
+            ZoneConstructionSets (list of ZoneConstructionSet): list of
+                ZoneConstructionSet objects.
+            ZoneLoads (list of ZoneLoad): list of ZoneLoad objects.
+            Zones (list of Zone): list of Zone objects
         """
         if Zones is None:
             Zones = []
@@ -139,117 +148,86 @@ class UmiTemplate:
         self.GlazingMaterials = GlazingMaterials
 
[…]
