### 2020-02-05 2c3481d
Fixes an issue where combining two schedules would use the schedule names in a dict
