### 2020-02-05 6e12568
Moves parallel_process to utils.py
