### 2020-02-05 d0886ec
Documents to use of parallel_process
