+++ docs/package_modules.rst
@@ -264,3 +264,4 @@ Utils
     lcm
     recursive_len
     rotate
+    parallel_process
