### 2020-02-05 samuelduchesne/archetypal#42 @samuelduchesne
Hi @brynpickering, as per 9c094567429d2a95bbeccbab9da2ecc18798f217, I believe this can be closed
