### 2020-02-05 samuelduchesne/archetypal#45 @samuelduchesne
Hi @brynpickering, as per your suggestion, I added skips for macOS builds (see #53). This pull request also added a macOS CI (using Travis). I believe this can be closed.
