### 2020-02-05 samuelduchesne/archetypal#43 @samuelduchesne
Hi @brynpickering, I think this can be closed.
