### 2020-02-05 samuelduchesne/archetypal#41 @samuelduchesne
Hi @ThibaultMarzullo, as per your recommendations, many improvements have been made to the paper (see #47 and c2823b38bbd1d2d8c72ef69f6fd0989a7da3dcaf ) and the documentation (see #62).

The documentation has been thoroughly upgraded with examples of workflows. All relevant packages and methods are also documented.
