### 2020-02-06 #1833 @samuelduchesne
> ...It looks like neither myself or @ThibaultMarzullo will be able to test the TRNSYS functionality, so some demonstration from your end might be necessary to convince us on that...?

Hi @ThibaultMarzullo and @brynpickering, regarding the TRNSYS functionality, I recorded the `archetypal convert` method that translates an IDF file to a TRNSYS model: https://youtu.be/jB5PhQ2YzKg.

The command has the `--trnsidf_exe` option pointing to the path of the executable which was located in the docker folder (the executable itself is not pushed to GitHub since it is a proprietary software). The two other arguments are the IDF file path and the weather file path. The output folder is not provided (which defaults to the current directory).
