### 2020-02-06 f1c8c4d
Adds readthedocs yml configuration file
