+++ .readthedocs.yml
@@ -26,5 +26,5 @@ python:
     - method: pip
       path: .
       extra_requirements:
-        - requirements-dev.txt
+          - dev
     system_packages: true
