+++ .readthedocs.yml
@@ -22,9 +22,8 @@ formats:
 python:
   version: 3.7
   install:
-    - requirements: requirements.txt
-    - method: pip
-      path: .
-      extra_requirements:
-          - dev
-    system_packages: true
+     - requirements: requirements.txt
+     - method: pip
+       path: .
+       extra_requirements:
+           - dev
