### 2020-02-07 f98fedb
Fixes an issue when the cli uses the current directory as a file path
