### 2020-02-11 samuelduchesne/archetypal#41 @ThibaultMarzullo
Hi @samuelduchesne and @louisleroy5 ,
I'm happy with the many improvements you did on the documentation and with the quality of the paper.

I have 2 additional suggestions for the paper. They are just small changes and English is not my first language, so I might be **very** wrong!

- In the "**Summary**" section: "_archetypal is a Python package that helps handle collections of such archetypes and to enable the interoperability [...]_", could it become "_archetypal is a Python package that helps **handling** collections of such archetypes and **enabling** the interoperability [...]_" ?

- In the "**E+ simulation environment**" section, end of paragraph 1: "_Other user-specified output names can also be specified._" Could it become "_Other output names can be specified by the user_" to remove repetition?

Regardless of my last 2 comments, this issue can be closed.
Thank you!
