### 2020-02-13 samuelduchesne/archetypal#45 @brynpickering
Great! Something to consider is moving CI to Azure pipelines, as it is currently better maintained than Travis (for reference, we recently moved to it in [our project](https://github.com/calliope-project/calliope/commit/4d7d5047a5c45d55aa937ffa6058579f4fb573e5)).
