### 2020-02-13 samuelduchesne/archetypal#43 @brynpickering
Looks good to me!
