### 2020-02-13 samuelduchesne/archetypal#42 @brynpickering
Seems to make sense from reading the file changes in the commit, but would be good to see online too. When's the planned next release?
