### 2020-02-13 samuelduchesne/archetypal#42 @samuelduchesne
oh! I forgot to say that the “develop” branch is also built on readthedocs.
![archetypal 1 3 0 — archetypal 1 3 0 documentation](https://user-images.githubusercontent.com/22966009/74442846-3870a000-4e40-11ea-832d-065f4fb34806.png)
