### 2020-02-17 samuelduchesne/archetypal#41 @samuelduchesne
Comments handled by f98fedb037dd9498a2112e2d81b50b537ab039e2
