### 2020-02-17 samuelduchesne/archetypal#46 @samuelduchesne
Hi @brynpickering, regarding this issue with `archetypal convert`, since you need TRNSYS for it work, please refer to https://github.com/openjournals/joss-reviews/issues/1833#issuecomment-582922979 for a demo.
