### 2020-02-17 #1833 @kyleniemeyer
Hi @brynpickering and @ThibaultMarzullo, it looks like @samuelduchesne has done some work in response to your review comments. Can you take another look now?
