### 2020-02-24 samuelduchesne/archetypal#63 @brynpickering
JOSS paper acronyms
Another JOSS review issue! Once again, just a minor one.

You should check your use of acronyms. `M` in `BEM` is given as `Models` but at one point is used in a way that means `modelling`. `UBEM` is given in full form twice in the document, once with the `M` referring to `Models` and once referring to `Modelling`. I would also not say `UBEM models` as it sounds weird when you think it means `Urban Building Energy Modelling models`.

In fact, my preference would be to avoid using both BEM and UBEM in your paper, just use the phrases in full form. It doesn't add much to the paper length, but does aid in clarity.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/1833)
