### 2020-02-24 #1833 @brynpickering
Hi @samuelduchesne, thanks for taking the time to address our issues and update archetypal. It looks to have changed markedly since the first set of reviews.
I've checked off most of the items on my review checklist, here are the outstanding items that I'd like some clarification on:

Installation
-----------
I'd like to re-do this, since I had issues the first time around. But it would be best on a stable release, which there doesn't seem to have been since July last year. Do you have an idea of your release schedule?

Performance
-------------
You talk about speed gain possibilities (inc. from caching) in your paper, but I don't see anything in the paper or the documentation to back up this claim.

I've also created [a minor issue](https://github.com/samuelduchesne/archetypal/issues/63) on the paper on your repo.
