### 2020-02-27 c398dcd
Fixes caching logic
