### 2020-02-28 591aaba
Better repr for EnergyPlusOptions
Also adds custom_processes in runargs
