### 2020-02-28 d6af7a9
Removes kwargs that do not impact caching from the hasher
