### 2020-03-02 #1833 @brynpickering
RE **versioning**: there's no reason to be on version 1.0.0 once it's ready for a public release; Pandas only released their version 1.0 this year, having reached v0.25! There's a lot of advice online about versioning (e.g. [here](https://semver.org/)) and I'm only an anecdotal source of wisdom on it. Nonetheless, my view on it is to only move to v1.0 once you're happy with the stability of the API; at the moment you can get away with backwards incompatible changes from one release to the next, and you don't necessarily need to be good about communicating them. Once you're >v1.0 you should have a relatively stable API and communicate clearly and in advance of upcoming breaking changes (Future/Deprecation warnings).

For the time being I can also just install archetypal from a local clone of your repository, if you'd prefer to release your next version only after the completion of the JOSS review.

RE **caching**: it does make sense. I would even add your detailed response to your documentation, to communicate it with potential users. What I was hoping for for this paper was some quantification of the performance increase. I.e. what is the difference in run-time thanks to caching. This could be mentioned in the paper and tabulated in the documentation. It should just be a matter of running the real world example you give here and recording time/memory consumption at each step.
