### 2020-03-02 698505f
Refactors zonesurfaces function with `exclude` parameter
method can exclude some surface types such as INTERNALMASS objects.
