+++ archetypal/template/building_template.py
@@ -425,11 +425,11 @@ class ZoneGraph(networkx.Graph):
             zone_obj = None
             if not skeleton:
                 zone_obj = Zone.from_zone_epbunch(zone, sql=sql)
-                zonesurfaces = zone.zonesurfaces
+                zonesurfaces = zone.zonesurfaces()
                 zone_obj._zonesurfaces = zonesurfaces
                 _is_core = zone_obj.is_core
             else:
-                zonesurfaces = zone.zonesurfaces
+                zonesurfaces = zone.zonesurfaces()
                 _is_core = is_core(zone)
             G.add_node(zone.Name, epbunch=zone, core=_is_core, zone=zone_obj)
 
@@ -585,7 +585,9 @@ class ZoneGraph(networkx.Graph):
 
             ggr = zone.theidf.idfobjects["GLOBALGEOMETRYRULES"][0]
 
-            for surface in zone.zonesurfaces:
+            for surface in zone.zonesurfaces(
+                exclude=["INTERNALMASS", "WINDOWSHADINGCONTROL"]
+            ):
                 if surface.key.lower() == "internalmass":
                     pass
                 else:
