### 2020-03-03 77c29de
Fixes issue #63
Streamlines the use if UBEM and BEM acronyms.
