### 2020-03-03 samuelduchesne/archetypal#63 @samuelduchesne
Hi @brynpickering, I've always struggled with these acronyms. You are right. I've removed most of them but I kept the first UBEM since is quite the buzzword in this field. ;)

Closing this
