### 2020-03-03 e1617c1
Addresses issue #65 (reading)
load_idf needs the `weather_file` parameter when passing in a weather file
