### 2020-03-03 5c668b6
Documentation reflects proper arguments and kwargs passed to eplus_run and load_eplus
Addresses issue #65
