### 2020-03-03 samuelduchesne/archetypal#65 @samuelduchesne
Hi @brynpickering; my bad! Reading an idf file in needs the `weather_file` keyword. Passing it as an argument won't work since the second argument is the `idd_filename`. I've fixed the documentation with 6fb4197c4084507ccba79a95ae09816d2fabda8f

```
idf2 = load_idf(eplus_file, weather_file=weather)
```

Same for the `run_eplus` method. weather_file must be passed as a keyword argument, not a positional argument.

I've made changes to the docs the reflect this issue. Let me know if the latest develop version works for you.
