+++ tests/test_template.py
@@ -1781,12 +1781,12 @@ class TestWindowSetting:
         idf, sql = small_idf
         idf2, sql2 = other_idf
         zone = idf.idfobjects["ZONE"][0]
-        iterator = iter([win for surf in zone.zonesurfaces() for win in
+        iterator = iter([win for surf in zone.zonesurfaces for win in
                          surf.subsurfaces])
         surface = next(iterator, None)
         window_1 = WindowSetting.from_surface(surface)
         zone = idf2.idfobjects["ZONE"][0]
-        iterator = iter([win for surf in zone.zonesurfaces() for win in
+        iterator = iter([win for surf in zone.zonesurfaces for win in
                          surf.subsurfaces])
         surface = next(iterator)
         window_2 = WindowSetting.from_surface(surface)
@@ -1805,13 +1805,13 @@ class TestWindowSetting:
         idf, sql = small_idf
         idf2, sql2 = other_idf
         zone = idf.idfobjects["ZONE"][0]
-        iterator = iter([win for surf in zone.zonesurfaces() for win in
+        iterator = iter([win for surf in zone.zonesurfaces for win in
                          surf.subsurfaces])
         surface = next(iterator, None)
         window_1 = WindowSetting.from_surface(surface)
         id_ = window_1.id
         zone = idf2.idfobjects["ZONE"][0]
-        iterator = iter([win for surf in zone.zonesurfaces() for win in
+        iterator = iter([win for surf in zone.zonesurfaces for win in
                          surf.subsurfaces])
         surface = next(iterator, None)
         window_2 = WindowSetting.from_surface(surface)
