### 2020-03-04 091040c
fixes calls to is_core
