### 2020-03-04 4263bd9
Accelerates tests by using cache (added config)
