### 2020-03-04 113f177
Simplifies tests using clean_config or fresh_start
fresh_start has been deleted and tests should use clean_fig instead
