### 2020-03-04 b935f21
Refactors tests so that they run quicker on CI
