+++ tests/test_trnsys.py
@@ -57,46 +57,49 @@ from archetypal.trnsys import (
     adds_sch_ground,
     adds_sch_setpoint,
 )
+from tests.conftest import get_platform
 
 
-class TestsConvert:
-    """Tests convert_idf_to_trnbuild()"""
+@pytest.fixture(
+    scope="class",
+    params=[
+        "RefBldgWarehouseNew2004_Chicago.idf",
+        "ASHRAE9012016_Warehouse_Denver.idf",
+        "ASHRAE9012016_ApartmentMidRise_Denver.idf",
+        "5ZoneGeometryTransform.idf",
+    ],
+)
+def converttest(self, request):
+    file = get_eplus_dirs(settings.ep_version) / "ExampleFiles" / request.param
+    window_file = "W74-lib.dat"
+    template_dir = os.path.join("archetypal", "ressources")
+    window_filepath = os.path.join(template_dir, window_file)
+    template_d18 = "tests/input_data/trnsys/NewFileTemplate.d18"
+    trnsidf_exe = "docker/trnsidf/trnsidf.exe"  # 'docker/trnsidf/trnsidf.exe'
 
-    @pytest.fixture(
-        scope="class",
-        params=[
-            "RefBldgWarehouseNew2004_Chicago.idf",
-            "ASHRAE9012016_Warehouse_Denver.idf",
-            "ASHRAE9012016_ApartmentMidRise_Denver.idf",
-            "5ZoneGeometryTransform.idf",
-        ],
+    # prepare args (key=value). Key is a unique id for the runs (here the
+    # file basename is used). Value is a dict of the function arguments
+    kwargs_dict = {
+        "u_value": 2.5,
+        "shgc": 0.6,
+        "t_vis": 0.78,
+        "tolerance": 0.05,
+        "ordered": True,
+    }
+    idf = load_idf(file)
+
+    weather_file = os.path.join(
+        "tests", "input_data", "CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw"
     )
-    def converttest(self, request):
-        file = get_eplus_dirs(settings.ep_version) / "ExampleFiles" / request.param
-        window_file = "W74-lib.dat"
-        template_dir = os.path.join("archetypal", "ressources")
-        window_filepath = os.path.join(template_dir, window_file)
-        template_d18 = "tests/input_data/trnsys/NewFileTemplate.d18"
-        trnsidf_exe = "docker/trnsidf/trnsidf.exe"  # 'docker/trnsidf/trnsidf.exe'
 
-        # prepare args (key=value). Key is a unique id for the runs (here the
-        # file basename is used). Value is a dict of the function arguments
-        kwargs_dict = {
-            "u_value": 2.5,
-            "shgc": 0.6,
-            "t_vis": 0.78,
-            "tolerance": 0.05,
-            "ordered": True,
-        }
-        idf = load_idf(file)
+    yield idf, file, weather_file, window_filepath, trnsidf_exe, template_d18, kwargs_dict
 
-        weather_file = os.path.join(
-            "tests", "input_data", "CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw"
-        )
+    del idf
 
-        yield idf, file, weather_file, window_filepath, trnsidf_exe, template_d18, kwargs_dict
 
-        del idf
+class TestConvert:
+
+    """Tests convert_idf_to_trnbuild()"""
 
     def test_get_save_write_schedules_as_input(self, config, converttest):
         output_folder = None
@@ -501,209 +504,168 @@ def trnbuild_file(config, request):
     yield idf_file
 
 
-@pytest.mark.xfail(
-    os.environ.get("CI", "False").lower() == "true",
[…]
