+++ tests/conftest.py
@@ -85,3 +85,20 @@ def pytest_runtest_setup(item):
 
 # dynamically define files to be ignored
 collect_ignore = ["test_core.py"]
+
+
+def get_platform():
+    """Returns the MacOS release number as tuple of ints"""
+    import platform
+
+    release, versioninfo, machine = platform.mac_ver()
+    release_split = release.split(".")
+    return tuple(map(safe_int_cast, release_split))
+
+
+def safe_int_cast(val, default=0):
+    """Safely casts a value to an int"""
+    try:
+        return int(val)
+    except (ValueError, TypeError):
+        return default
