### 2020-03-04 5f98f36
Fixes test outside class
