### 2020-03-04 ac4e520
Adds timeit to cli and only tests one file
