### 2020-03-04 8e96024
Simplifies naming of duplicated objects
+ Better docstrings
