### 2020-03-06 samuelduchesne/archetypal#65 @brynpickering
Yep, that all runs smoothly now. Thanks!
