### 2020-03-13 e91f96b
Creates an error file when reduce fails for some idf files
