### 2020-03-16 919adc4
Fixes an issue with dict filter
