### 2020-03-16 bb12993
Fixes errors file list creation
