### 2020-03-25 f110297
Refactors dataportal module with more streamlined code
