### 2020-03-25 515bd0f
Fixes an issue with cli test command arguments
They are not the same as in the cli methods. This was not detected on travis CI because they are skipped.
