+++ tests/test_cli.py
@@ -72,7 +72,7 @@ class TestCli:
                 "convert",
                 "--ep_version",
                 "9-2-0",
-                "--window-lib",
+                "--window_lib",
                 "archetypal/ressources/W74-lib.dat",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
@@ -116,7 +116,7 @@ class TestCli:
                 "convert",
                 "--ep_version",
                 "9-2-0",
-                "--log-clear-names",
+                "--log_clear_names",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
                 "--trnsidf_exe",
@@ -212,7 +212,7 @@ class TestCli:
                 "convert",
                 "--ep_version",
                 "9-2-0",
-                "--schedule-as-input",
+                "--schedule_as_input",
                 "tests/input_data/trnsys/simple_2_zone.idf",
                 "tests/input_data/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw",
                 "--trnsidf_exe",
