### 2020-03-25 1f7857d
Should return the json_response instead of the whole response object
