+++ tests/test_trnsys.py
@@ -575,7 +575,7 @@ class TestTrnBuild:
     def test_trnbuild_from_idf_parallel_darwin_or_linux(self, config):
         # All IDF files
         # List files here
-        file_upper_path = os.path.join(settings.ep_version, "ExampleFiles")
+        file_upper_path = os.path.join(get_eplus_dirs(settings.ep_version), "ExampleFiles")
         files = [
             "RefBldgWarehouseNew2004_Chicago.idf",
             "ASHRAE9012016_Warehouse_Denver.idf",
