### 2020-03-26 bd6774a
Adds note to install git before using git cmd
