### 2020-03-26 c40e4d5
Adds meteo file path to examples
