### 2020-03-27 dbe2ac4
Adds logic for detecting dimming types
