+++ archetypal/template/load.py
@@ -10,6 +10,7 @@ import collections
 from archetypal import log, timeit, settings
 from archetypal.template import UmiBase, Unique, UmiSchedule, UniqueName
 from archetypal.utils import reduce
+import logging as lg
 
 
 class ZoneLoad(UmiBase, metaclass=Unique):
@@ -36,7 +37,7 @@ class ZoneLoad(UmiBase, metaclass=Unique):
         IsLightingOn=True,
         IsPeopleOn=True,
         PeopleDensity=0.2,
-        **kwargs
+        **kwargs,
     ):
         """Initialize a new ZoneLoad object
 
@@ -180,8 +181,8 @@ class ZoneLoad(UmiBase, metaclass=Unique):
             zone (archetypal.template.zone.Zone): zone to gets information from
         """
 
-        # Get schedule index for different loads and creates ZoneLoad arguments
-        # Verifies if Equipment in zone
+        # Get schedule index for different loads and create ZoneLoad arguments
+        # Verify if Equipment in zone
         zone_index = zone.sql["Zones"][
             zone.sql["Zones"]["ZoneName"].str.contains(zone.Name.upper())
         ].index[0]
@@ -202,7 +203,7 @@ class ZoneLoad(UmiBase, metaclass=Unique):
             else:
                 EquipmentPowerDensity = (
                     nominal_elec["DesignLevel"].sum() + nominal_gas["DesignLevel"].sum()
-                ) / zone.area
+                ) / zone.area  # todo: Should nominal gas really be added to elec?
 
             sched_indexes = nominal_elec["ScheduleIndex"].values
             design_index = nominal_elec["DesignLevel"].index
@@ -210,11 +211,11 @@ class ZoneLoad(UmiBase, metaclass=Unique):
             for sched, design in zip(sched_indexes, design_index):
                 sched_name = zone.sql["Schedules"]["ScheduleName"][sched]
                 schedule = UmiSchedule(Name=sched_name, idf=zone.idf)
-                schedule.weights = nominal_elec["DesignLevel"][design]
+                schedule.combine_weight = nominal_elec["DesignLevel"][design]
                 list_sched.append(schedule)
 
             EquipmentAvailabilitySchedule = reduce(
-                UmiSchedule.combine, list_sched, weights="weights"
+                UmiSchedule.combine, list_sched, weights="combine_weight"
             )
         # Verifies if Lights in zone
         if zone.sql["NominalLighting"][
@@ -265,10 +266,10 @@ class ZoneLoad(UmiBase, metaclass=Unique):
         z_load = cls(
             Name=name,
             zone=zone,
-            DimmingType="Continuous",
+            DimmingType=_resolve_dimming_type(zone),
             EquipmentAvailabilitySchedule=EquipmentAvailabilitySchedule,
             EquipmentPowerDensity=EquipmentPowerDensity,
-            IlluminanceTarget=500,
+            IlluminanceTarget=_resolve_illuminance_target(zone),
             LightingPowerDensity=LightingPowerDensity,
             LightsAvailabilitySchedule=LightsAvailabilitySchedule,
             OccupancySchedule=OccupancySchedule,
@@ -359,3 +360,85 @@ class ZoneLoad(UmiBase, metaclass=Unique):
         new_obj._belongs_to_zone = self._belongs_to_zone
         new_obj._predecessors.extend(self.predecessors + other.predecessors)
         return new_obj
+
+
+def _resolve_dimming_type(zone):
+    """Resolves the dimming type for the Zone object"""
+    # First, retrieve the list of Daylighting objects for this zone. Uses the eppy
+    # `getreferingobjs` method.
+    ep_obj = zone._epbunch
[…]
