### 2020-03-27 ae57d9a
Completes the docstring
