### 2020-03-27 e514b71
Switches to numpy equality test
hopefully, this is quicker
