### 2020-03-27 984b5da
Adds weights list length test + more comments
