### 2020-03-27 aa68649
Removes unecessary *args parameter in constructor
