### 2020-03-27 8f7c8cc
Improved docstring for class ZoneLoad
