+++ archetypal/template/load.py
@@ -41,31 +41,25 @@ class ZoneLoad(UmiBase, metaclass=Unique):
         """Initialize a new ZoneLoad object
 
         Args:
-            *args:
-            DimmingType (str): Different types to dim the lighting to respect
-                the IlluminanceTraget and taking into account the daylight
-                illuminance: * If `Continuous` : the overhead lights dim
-                continuously and
-
-                    linearly from (maximum electric power, maximum light output)
-                    to (minimum electric power, minimum light output) as the
-                    daylight illuminance increases. The lights stay on at the
-                    minimum point with further increase in the daylight
-                    illuminance
-
-                * If `Stepped`: the electric power input and light output vary
-                      in discrete, equally spaced steps
-
-                * If `Off`: Lights switch off completely when the minimum
-                      dimming point is reached
+            DimmingType (str): Different types to dim the lighting to respect the
+                IlluminanceTraget and taking into account the daylight illuminance:
+                    - If `Continuous`, the overhead lights dim continuously and linearly
+                      from (maximum electric power, maximum light output) to (minimum
+                      electric power, minimum light output) as the daylight illuminance
+                      increases. The lights stay on at the minimum point with further
+                      increase in the daylight illuminance.
+                    - If `Stepped`, the electric power input and light output vary
+                      in discrete, equally spaced steps.
+                    - If `Off`, Lights switch off completely when the minimum
+                      dimming point is reached.
             EquipmentAvailabilitySchedule (UmiSchedule): The name of
                 the schedule (Day | Week | Year) that modifies the design level
                 parameter for electric equipment.
             EquipmentPowerDensity (float): Equipment Power Density in the zone
-                (W/m²)
+                (W/m²).
             IlluminanceTarget (float): Number of lux to be respected in the zone
             LightingPowerDensity (float): Lighting Power Density in the zone
-                (W/m²)
+                (W/m²).
             LightsAvailabilitySchedule (UmiSchedule): The name of the
                 schedule (Day | Week | Year) that modifies the design level
                 parameter for lighting.
@@ -73,13 +67,13 @@ class ZoneLoad(UmiBase, metaclass=Unique):
                 (Day | Week | Year) that modifies the number of people parameter
                 for electric equipment.
             IsEquipmentOn (bool): If True, heat gains from Equipment are taken
-                into account for the zone's load calculation
+                into account for the zone's load calculation.
             IsLightingOn (bool): If True, heat gains from Lights are taken into
-                account for the zone's load calculation
+                account for the zone's load calculation.
             IsPeopleOn (bool): If True, heat gains from People are taken into
-                account for the zone's load calculation
-            PeopleDensity (float): Density of people in the zone (people/m²)
-            **kwargs:
+                account for the zone's load calculation.
+            PeopleDensity (float): Density of people in the zone (people/m²).
+            **kwargs: Other keywords passed to the parent constructor :class:`UmiBase`.
         """
         super(ZoneLoad, self).__init__(**kwargs)
         self.DimmingType = DimmingType
