### 2020-03-27 f10875c
Fixes the default illuminance target
