### 2020-03-27 6c1fe29
Adds tests for new dimming type and illuminance target logic
