+++ tests/test_template.py
@@ -1154,6 +1154,28 @@ class TestZoneConstructionSet:
 class TestZoneLoad:
     """Combines different :class:`ZoneLoad` tests"""
 
+    @pytest.fixture(scope="class")
+    def fiveZoneEndUses(self, config):
+        file = (
+            get_eplus_dirs(settings.ep_version)
+            / "ExampleFiles"
+            / "5ZoneAirCooled_AirBoundaries_Daylighting.idf"
+        )
+        w = (
+            get_eplus_dirs(settings.ep_version)
+            / "WeatherData"
+            / "USA_IL_Chicago-OHare.Intl.AP.725300_TMY3.epw"
+        )
+        idf = ar.load_idf(file, weather_file=w)
+        sql = idf.run_eplus(
+            output_report="sql",
+            prep_outputs=True,
+            annual=False,
+            design_day=True,
+            verbose="v",
+        )
+        yield idf, sql
+
     @pytest.fixture(scope="class", params=["RefBldgWarehouseNew2004_Chicago.idf"])
     def zoneLoadtests(self, config, request):
         """
@@ -1172,7 +1194,7 @@ class TestZoneLoad:
             output_report="sql",
             prep_outputs=True,
             annual=False,
-            design_day=False,
+            design_day=True,
             verbose="v",
         )
         yield idf, sql
@@ -1197,7 +1219,35 @@ class TestZoneLoad:
         idf, sql = zoneLoadtests
         zone = idf.getobject("ZONE", "Office")
         z = Zone.from_zone_epbunch(zone_ep=zone, sql=sql)
-        load_ = ZoneLoad.from_zone(z)
+        zone_loads = ZoneLoad.from_zone(z)
+
+        assert zone_loads.DimmingType == "Off"
+        assert zone_loads.EquipmentPowerDensity == 8.07
+        assert zone_loads.IlluminanceTarget == 500
+        assert zone_loads.IsEquipmentOn
+        assert zone_loads.IsPeopleOn
+        assert zone_loads.LightingPowerDensity == 11.84
+        np.testing.assert_almost_equal(zone_loads.PeopleDensity, 0.021107919980354082)
+
+    def test_zoneLoad_from_zone_mixedparams(self, config, fiveZoneEndUses):
+        from archetypal import ZoneLoad, Zone
+
+        idf, sql = fiveZoneEndUses
+        zone = idf.getobject("ZONE", "SPACE1-1")
+        z = Zone.from_zone_epbunch(zone_ep=zone, sql=sql)
+        zone_loads = ZoneLoad.from_zone(z)
+
+        assert zone_loads.DimmingType == "Stepped"
+        np.testing.assert_almost_equal(
+            zone_loads.EquipmentPowerDensity, 10.649455425574827
+        )
+        assert zone_loads.IlluminanceTarget == 400
+        assert zone_loads.IsEquipmentOn
+        assert zone_loads.IsPeopleOn
+        np.testing.assert_almost_equal(
+            zone_loads.LightingPowerDensity, 15.974, decimal=3
+        )
+        np.testing.assert_almost_equal(zone_loads.PeopleDensity, 0.111, decimal=3)
 
     def test_zoneLoad_from_to_json(self, config):
         """
@@ -1313,7 +1363,7 @@ class TestZoneConditioning:
             output_report="sql",
             prep_outputs=True,
             annual=False,
-            design_day=False,
+            design_day=True,
             verbose="v",
         )
         yield idf, sql, request.param
