### 2020-03-28 82623b4
Adds the schedule report to SQL database
