### 2020-03-28 fa12d0d
Initial refactored/Improved logic for dealing with mechanical ventilation
