+++ archetypal/template/conditioning.py
@@ -11,6 +11,7 @@ import math
 import numpy as np
 from archetypal import float_round, ReportData, log, timeit, settings
 from archetypal.template import UmiBase, Unique, UmiSchedule, UniqueName
+import logging as lg
 
 
 class ZoneConditioning(UmiBase, metaclass=Unique):
@@ -341,69 +342,126 @@ class ZoneConditioning(UmiBase, metaclass=Unique):
         """Iterate on 'Controller:MechanicalVentilation' objects to find the
         'DesignSpecifactionOutdoorAirName' for the zone.
 
-        Todo:
-            - Get mechanical sizing by zone.
-
         Args:
             zone (Zone): The zone object.
         """
-        if not zone.idf.idfobjects["Controller:MechanicalVentilation".upper()]:
-            IsMechVentOn = False
-            MinFreshAirPerArea = 0  # use defaults
-            MinFreshAirPerPerson = 0
-            MechVentSchedule = UmiSchedule.constant_schedule(idf=zone.idf)
-        else:
-            design_spe_outdoor_air_name = ""
-            for object in zone.idf.idfobjects[
-                "Controller:MechanicalVentilation".upper()
-            ]:
-                if zone.Name in object.fieldvalues:
-                    indice_zone = [
-                        k for k, s in enumerate(object.fieldvalues) if s == zone.Name
-                    ][0]
-                    design_spe_outdoor_air_name = object.fieldvalues[indice_zone + 1]
+        # For models with ZoneSizes
+        try:
+            try:
+                (
+                    self.IsMechVentOn,
+                    self.MinFreshAirPerArea,
+                    self.MinFreshAirPerPerson,
+                    self.MechVentSchedule,
+                ) = self.fresh_air_from_zone_sizes(zone)
+            except ValueError:
+                (
+                    self.IsMechVentOn,
+                    self.MinFreshAirPerArea,
+                    self.MinFreshAirPerPerson,
+                    self.MechVentSchedule,
+                ) = self.fresh_air_from_ideal_loads(zone)
+        except:
+            self.IsMechVentOn = False
+            self.MinFreshAirPerPerson = 0
+            self.MinFreshAirPerArea = 0
+            self.MechVentSchedule = UmiSchedule.constant_schedule(idf=zone.idf)
 
-                    if object.Availability_Schedule_Name != "":
-                        MechVentSchedule = UmiSchedule(
-                            Name=object.Availability_Schedule_Name, idf=zone.idf
-                        )
-                    else:
-                        MechVentSchedule = UmiSchedule.constant_schedule(idf=zone.idf)
-                    break
-            # If 'DesignSpecifactionOutdoorAirName', MechVent is ON, and gets
-            # the minimum fresh air (per person and area)
-            if design_spe_outdoor_air_name != "":
-                IsMechVentOn = True
-                design_spe_outdoor_air = zone.idf.getobject(
-                    "DesignSpecification:OutdoorAir".upper(),
-                    design_spe_outdoor_air_name,
-                )
-                MinFreshAirPerPerson = (
-                    design_spe_outdoor_air.Outdoor_Air_Flow_per_Person
-                )
-                # If MinFreshAirPerPerson NOT a number, MinFreshAirPerPerson=0
-                try:
-                    MinFreshAirPerPerson = float(MinFreshAirPerPerson)
-                except:
-                    MinFreshAirPerPerson = 0
-                MinFreshAirPerArea = (
-                    design_spe_outdoor_air.Outdoor_Air_Flow_per_Zone_Floor_Area
-                )
[…]
