### 2020-03-28 aca26c9
To Static Methods
