### 2020-04-15 102a133
Fixes ref to networkx docs
