### 2020-04-15 f596120
Fixes alignment of text for bullet lists
