### 2020-04-15 5523b23
Sets up stages
