+++ .travis.yml
@@ -12,8 +12,22 @@ jobs:
   include:
   - os: linux
     env: TRAVIS_PYTHON_VERSION=3.7
+    script: py.test --cov=archetypal --verbose tests/
   - os: osx
     env: TRAVIS_PYTHON_VERSION=3.7
+    script: py.test --cov=archetypal --verbose tests/
+  - stage: deploy
+    script: echo "Deploying to GitHub releases ..."
+    deploy:
+      - provider: pypi
+        user: samueld
+        password:
+          secure: ULsmo10liyAUeo3zh6gz8ORuYsg39kMt07uGz4x8YHAAg/d21tXv23H3R7NKUKqoNlcPPPlQXNOZD7lrKvpUSdHlth8xTunrABxjYNcvaFD0eQolV3N3XPP+ePHftNCz3hm65aKh014DdwueWISFsOun0S16jlhJkak7yRJ5zWEKC/9kIhh/iH6DToMWCLc/o5wNT7b7G2OKvojXbotkyh1GiTgSgIVHYEOPVlKUsUsifvu9wvWuf2u3ymbEKLDTKoUDyjaao8Gs1H7Wms0BB0UjjQquJnbwJjB5bx4gDp7HytoLOdzanfdt3zykgx6K5+dq0533vOUTCFWeedLknwt6HUrqXSGOYQxQwoTmHjDrmpgG+Rw8zHoGhCzGtQ0co6JYrSpdnbjv98xJkPKKanam/WkZ+SQIzDgBzFn4S4fa9hJsM/qkgxuY9r4FUA5WFaMRwOQ3NW52hj9WYJ7bjDRmh9XjAbHnyYvgS1B/9Q21GurgT9B1C9uKQ70vzKmIC8jrYeSPTOVNE4MN3hptFjIgHVFP5Va/xcLSbkFnX1oDmCMm0eKaemjABBMboy9k2Wc+rPGhCksyWom4v7Mm6i042MB7IEowCC4toYvkoXZCOtdKjzG4nsKccTf2xBYM+f8NT4H/TIfRI60NE8wKTEHzq2YaTRHYX2aYMYA4JKc=
+        on:
+          tags: true
+          distributions: sdist bdist_wheel
+          repo: samuelduchesne/archetypal
+          branch: release/v1.2.7
 before_install:
 - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then EXT=dmg; PLATFORM=Darwin; fi
 - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then EXT=sh; PLATFORM=Linux; fi
@@ -58,18 +72,6 @@ install:
 - pip install --upgrade setuptools
 - pip install --upgrade pip
 - pip install .[dev]
-script:
-#- py.test --cov=archetypal --verbose tests/
 after_success:
 - coverage report -m
 - coveralls
-deploy:
-  - provider: pypi
-    user: samueld
-    password:
[…]
