### 2020-04-15 25e58ec
Organizes builds + adds Github Releases deploy
