+++ .travis.yml
@@ -2,32 +2,41 @@ language: c
 env:
   global:
   - ARCHETYPAL_INTEGRATION=True ENERGYPLUS_VERSION=9.2.0 ENERGYPLUS_SHA=921312fa1d
-    ENERGYPLUS_INSTALL_VERSION=9-2-0 MINICONDA_VERSION=latest
+    ENERGYPLUS_INSTALL_VERSION=9-2-0 MINICONDA_VERSION=latest TRAVIS_PYTHON_VERSION=3.7
 addons:
   apt:
     packages:
     - wine
     - libgfortran3
+os:
+- osx
+- linux
 jobs:
   include:
-  - os: linux
-    env: TRAVIS_PYTHON_VERSION=3.7
-    script: echo "Running Tests ..." # py.test --cov=archetypal --verbose tests/
-  - os: osx
-    env: TRAVIS_PYTHON_VERSION=3.7
-    script: echo "Running Tests ..." #py.test --cov=archetypal --verbose tests/
+  - stage: test
+    script: echo "Running Tests ..."
+    if: branch != release/v1.2.7
   - stage: deploy
     script: echo "Deploying to GitHub releases ..."
     deploy:
-      - provider: pypi
-        user: samueld
-        password:
-          secure: ULsmo10liyAUeo3zh6gz8ORuYsg39kMt07uGz4x8YHAAg/d21tXv23H3R7NKUKqoNlcPPPlQXNOZD7lrKvpUSdHlth8xTunrABxjYNcvaFD0eQolV3N3XPP+ePHftNCz3hm65aKh014DdwueWISFsOun0S16jlhJkak7yRJ5zWEKC/9kIhh/iH6DToMWCLc/o5wNT7b7G2OKvojXbotkyh1GiTgSgIVHYEOPVlKUsUsifvu9wvWuf2u3ymbEKLDTKoUDyjaao8Gs1H7Wms0BB0UjjQquJnbwJjB5bx4gDp7HytoLOdzanfdt3zykgx6K5+dq0533vOUTCFWeedLknwt6HUrqXSGOYQxQwoTmHjDrmpgG+Rw8zHoGhCzGtQ0co6JYrSpdnbjv98xJkPKKanam/WkZ+SQIzDgBzFn4S4fa9hJsM/qkgxuY9r4FUA5WFaMRwOQ3NW52hj9WYJ7bjDRmh9XjAbHnyYvgS1B/9Q21GurgT9B1C9uKQ70vzKmIC8jrYeSPTOVNE4MN3hptFjIgHVFP5Va/xcLSbkFnX1oDmCMm0eKaemjABBMboy9k2Wc+rPGhCksyWom4v7Mm6i042MB7IEowCC4toYvkoXZCOtdKjzG4nsKccTf2xBYM+f8NT4H/TIfRI60NE8wKTEHzq2YaTRHYX2aYMYA4JKc=
-        on:
-          tags: true
-          distributions: sdist bdist_wheel
-          repo: samuelduchesne/archetypal
-          branch: release/v1.2.7
+    - provider: pypi
+      user: samueld
+      password:
[…]
