+++ .travis.yml
@@ -15,9 +15,8 @@ jobs:
   include:
   - stage: test
     script: echo "Running Tests ..."
-    if: branch != release/v1.2.7
   - stage: deploy
-    script: echo "Deploying to GitHub releases ..."
+    script: echo "Deploying to PyPi and GitHub releases ..."
     deploy:
     - provider: pypi
       user: samueld
@@ -33,10 +32,11 @@ jobs:
       api_key:
         secure: T0s3DtR5Gqnw7tMIC82Pn/pqWYwF5IJ1sppX2E+VqgJnZ37g59qFDK2g53e4PgXqkoNb5VhO54wYu/v8pms8LOtH9DYq/OsERAw0vr/qlJRIe78qFI+RaWdLS6lXE4K3fN7au+yqSpxmWmeZFq0k9BkiX7tWcNxsykcVCsI/MYziMbJD6mku1KSr9+/dYQO/XqARDOK8+faPE0jJrE6aBBMJa27c6jmoZsvnsgRL4tx73f8hsW2HXh2mmpBvsBcKi0FG9XwSjmLF/vL4pad0rF2XSdsCHwoVHzFfdUIi3Iu6FKsUTy6pxtOA/TMPdJ6xRzP/snURX1UAAIRWy+hg+bt2VsGcsCndWY78DGKHO3xolRIdc6JoOh+qRwcQyJfJVC1MjTyFhyhdgpZzhdqDp4bomatwGEyWtRU6L7MDIbhv3wiui28Jdfyw++ftxVa+JqOJRJ0MNDF9XBa+cze/os4MvHcrCFg7HBasdnqUSVz/exmPVBVsYIQWcgOQGbAFguXgCOf6N02wg1NHW9qWGbk+MMgv8sE3spBMZAnY/9Ll5dfXYHwEipnLYRIAN87fR64fKtkdY6/rxQMAY/qQzvPO9WSPHGOMDanywryctX5Ki+L0H8fXMpuwZhYMb85sJlEiIUo2/LCmD76gCqGj+pt8jigH2y9W76XU6EENEjc=
       file_glob: true
-      file: "dist/archetypal*"
+      file: dist/archetypal*
       on:
         repo: samuelduchesne/archetypal
         branch: release/v1.2.7
+      skip_cleanup: true
 before_install:
 - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then EXT=dmg; PLATFORM=Darwin; fi
 - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then EXT=sh; PLATFORM=Linux; fi
