+++ .travis.yml
@@ -86,9 +86,9 @@ before_install:
 - conda env update -q -n test-env python=$TRAVIS_PYTHON_VERSION -f environment.yml
 - conda activate test-env
 install:
-- pip install --upgrade setuptools
-- pip install --upgrade pip
-- pip install .[dev]
+- pip3 install --upgrade pip
+- pip3 install --upgrade setuptools
+- pip3 install .[dev]
 after_success:
 - coverage report -m
 - coveralls
