+++ .travis.yml
@@ -10,20 +10,20 @@ addons:
     - libgfortran3
 jobs:
   include:
-    - name: Python 3.7 on Xenial Linux
-      python: 3.7           # this works for Linux but is ignored on macOS or Windows
-    - name: Python 3.7.5 on macOS
+    - name: Python 3.8 on Xenial Linux
+      python: 3.8           # this works for Linux but is ignored on macOS or Windows
+    - name: Python 3.8 on macOS
       os: osx
-      osx_image: xcode11.3
+      osx_image: xcode11.2
       language: objective-c
-      env: PYENV_VERSION=3.7.5
+      env: PYENV_VERSION=3.8
     - name: "Python 3.7 on Windows"
       os: windows           # Windows 10.0.17134 N/A Build 17134
       language: shell       # 'language: python' is an error on Travis CI Windows
       before_install:
-        - choco install python --version 3.7.4
+        - choco install python --version 3.8
         - python -m pip install --upgrade pip
-      env: PATH=/c/Python37:/c/Python37/Scripts:$PATH
+      env: PATH=/c/Python38:/c/Python38/Scripts:$PATH
     - stage: deploy
       script: echo "Deploying to PyPi and GitHub releases ..."
       deploy:
@@ -91,7 +91,7 @@ before_install:
 install:
 - pip3 install --upgrade pip
 - pip3 install --upgrade setuptools
-- pip3 install .[dev]
+- pip3 install --user .[dev]
 after_success:
 - coverage report -m
 - coveralls
