+++ .travis.yml
@@ -81,8 +81,8 @@ before_install:
   fi
 - wget https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.sh
   -O miniconda.sh
-- bash miniconda.sh -b -p "$HOME"/miniconda
-- source "$HOME"/miniconda/etc/profile.d/conda.sh
+- bash miniconda.sh -b -p $HOME/miniconda
+- source $HOME/miniconda/etc/profile.d/conda.sh
 - conda config --set always_yes yes --set changeps1 no
 - conda update -q conda;
 - conda info -a
