### 2020-04-15 9e39c60
Moves python before_install in global before_install
