### 2020-04-15 951f506
Setting energyplus install for travis windows
