+++ .travis.yml
@@ -63,21 +63,23 @@ before_install:
   echo "y\r" | sudo ./$ENERGYPLUS_DOWNLOAD_FILENAME.$EXT; sudo tar zxvf $ATTCHNUM.zip
   -C /usr/local/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/PreProcess/IDFVersionUpdater;
   sudo chmod -R a+rwx /usr/local/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/PreProcess/IDFVersionUpdater;
-  sudo chmod -R a+rwx /usr/local/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/ExampleFiles;
+  sudo chmod -R a+rwx /usr/local/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/ExampleFiles; sudo rm
+  $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT;
+  sudo rm $ATTCHNUM.zip;
   fi
 - if [ "$TRAVIS_OS_NAME" == "osx" ]; then curl -SLO https://raw.githubusercontent.com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs;
   sudo hdiutil attach $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT; sudo /Volumes/$ENERGYPLUS_DOWNLOAD_FILENAME/$ENERGYPLUS_DOWNLOAD_FILENAME.app/Contents/MacOS/$ENERGYPLUS_DOWNLOAD_FILENAME
   --verbose --script install_script.qs; sudo tar zxvf $ATTCHNUM.zip -C /Applications/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/PreProcess;
   sudo chmod -R a+rwx /Applications/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/PreProcess/IDFVersionUpdater;
-  sudo chmod -R a+rwx /Applications/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/ExampleFiles;
+  sudo chmod -R a+rwx /Applications/EnergyPlus-$ENERGYPLUS_INSTALL_VERSION/ExampleFiles; sudo rm
+  $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT;
+  sudo rm $ATTCHNUM.zip;
   fi
 - if [ "$TRAVIS_OS_NAME" == "windows" ]; then curl -SL %EXTRAS_DOWNLOAD_URL% -o %ATTCHNUM%.zip; SET
   DEST=C:\EnergyPlusV%ENERGYPLUS_INSTALL_VERSION%\PreProcess\IDFVersionUpdater; ECHO Extracting and Copying files to.
   ..; ECHO %DEST%; ATTCHNUM = 8231; DEST =
   "C:\EnergyPlusV$env:ENERGYPLUS_INSTALL_VERSION\PreProcess\IDFVersionUpdater"; ps: "Expand-Archive -Path
   C:\\projects\\archetypal\\$env:ATTCHNUM.zip -DestinationPath $env:DEST -Force"; DEL /s "%ATTCHNUM%.zip"; fi
-- sudo rm $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT
-- sudo rm $ATTCHNUM.zip
 - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then MINICONDA_OS=Linux; fi
 - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then MINICONDA_OS=MacOSX; fi
 - if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then MINICONDA_OS=Windows; fi
