+++ .travis.yml
@@ -75,12 +75,11 @@ before_install:
   $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT;
   sudo rm $ATTCHNUM.zip;
   fi
-- if [ "$TRAVIS_OS_NAME" == "windows" ]; then curl -SL $env:EXTRAS_DOWNLOAD_URL -o $env:ATTCHNUM.zip; SET
-  DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; ECHO Extracting and Copying files to.
-  ..; ECHO $env:DEST; ATTCHNUM = 8231; DEST=C:\\EnergyPlusV$env:ENERGYPLUS_INSTALL_VERSION\\PreProcess
-  \\IDFVersionUpdater; fi
-- if [ "$TRAVIS_OS_NAME" == "windows" ] then Expand-Archive -Path C:\\projects\\archetypal\\$env:ATTCHNUM.zip
-  -DestinationPath $env:DEST -Force; DEL /s "$env:ATTCHNUM.zip"; fi
+- if [ "$TRAVIS_OS_NAME" == "windows" ]; then
+  DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; ECHO Extracting and Copying files to
+  ...; ECHO $env:DEST; DEST=C:\\EnergyPlusV$env:ENERGYPLUS_INSTALL_VERSION\\PreProcess\\IDFVersionUpdater;
+  powershell Expand-Archive -Path C:\\projects\\archetypal\\$env:ATTCHNUM.zip -DestinationPath $env:DEST -Force; DEL /s
+  "$env:ATTCHNUM.zip"; fi
 - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then MINICONDA_OS=Linux; fi
 - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then MINICONDA_OS=MacOSX; fi
 - if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then MINICONDA_OS=Windows; fi
