### 2020-04-15 231958d
Actually installs E+ on windows
