+++ .travis.yml
@@ -77,10 +77,9 @@ before_install:
   fi
 - if [ "$TRAVIS_OS_NAME" == "windows" ]; then curl -SLO https://raw.githubusercontent
   .com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs;
-  $ENERGYPLUS_DOWNLOAD_FILENAME --verbose --script install_script.qs; fi
+  $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs; fi
 - if [ "$TRAVIS_OS_NAME" == "windows" ]; then
-  DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater;
-  DEST=C:\\EnergyPlusV$env:ENERGYPLUS_INSTALL_VERSION\\PreProcess\\IDFVersionUpdater; echo "Extracting and Copying files to
+  DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $env:DEST;
   powershell -command "Expand-Archive -Path C:\\projects\\archetypal\\$env:ATTCHNUM.zip -DestinationPath $env:DEST
   -Force"; DEL /s
