### 2020-04-15 #1833 @brynpickering
Would appreciate seeing an update to the paper / docs on cache performance, otherwise happy with everything 👍
