+++ .travis.yml
@@ -75,8 +75,7 @@ before_install:
   $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT;
   sudo rm $ATTCHNUM.zip;
   fi
-- if [ "$TRAVIS_OS_NAME" == "windows" ]; then curl -SLO https://raw.githubusercontent
-  .com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs;
+- if [ "$TRAVIS_OS_NAME" == "windows" ]; then curl -0 https://raw.githubusercontent.com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs -o install_script.qs;
   $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs; fi
 - if [ "$TRAVIS_OS_NAME" == "windows" ]; then
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
