### 2020-04-15 0ff5f35
Moves python install after E+ install
