+++ .travis.yml
@@ -76,8 +76,7 @@ before_install:
   sudo rm $ATTCHNUM.zip;
   fi
 - if [ "$TRAVIS_OS_NAME" == "windows" ]; then curl -0 https://raw.githubusercontent.com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs -o install_script.qs;
-  $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs; fi
-- if [ "$TRAVIS_OS_NAME" == "windows" ]; then
+  $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs; dir;
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $env:DEST;
   powershell -command "Expand-Archive -Path C:\\projects\\archetypal\\$env:ATTCHNUM.zip -DestinationPath $env:DEST
