### 2020-04-15 214519e
removes env before paths
