+++ .travis.yml
@@ -79,7 +79,7 @@ before_install:
   $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs; dir;
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $DEST;
-  powershell -command Expand-Archive -Path C:\projects\archetypal\$ATTCHNUM.zip -DestinationPath $DEST -Force; fi
+  powershell -command Expand-Archive -Path $ATTCHNUM.zip -DestinationPath $DEST -Force; fi
 - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then MINICONDA_OS=Linux; fi
 - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then MINICONDA_OS=MacOSX; fi
 - if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then MINICONDA_OS=Windows; fi
