### 2020-04-15 173b16c
Simplifies env variable + uses curl instead of wget
