+++ .travis.yml
@@ -12,16 +12,16 @@ jobs:
   include:
     - name: Python 3.8 on Xenial Linux
       python: 3.8           # this works for Linux but is ignored on macOS or Windows
-      env: MINICONDA_OS=Linux
+      env: MINICONDA_OS=Linux; MINICONDAEXT=sh
     - name: Python 3.8 on macOS
       os: osx
       osx_image: xcode11.2
       language: objective-c
-      env: PYENV_VERSION=3.8; MINICONDA_OS=MacOSX
+      env: PYENV_VERSION=3.8; MINICONDA_OS=MacOSX; MINICONDAEXT=sh
     - name: "Python 3.8 on Windows"
       os: windows           # Windows 10.0.17134 N/A Build 17134
       language: shell       # 'language: python' is an error on Travis CI Windows
-      env: PATH=/c/Python38:/c/Python38/Scripts:$PATH; MINICONDA_OS=Windows
+      env: PATH=/c/Python38:/c/Python38/Scripts:$PATH; MINICONDA_OS=Windows; MINICONDAEXT=exe
     - stage: deploy
       script: echo "Deploying to PyPi and GitHub releases ..."
       deploy:
@@ -77,12 +77,14 @@ before_install:
   sudo rm $ATTCHNUM.zip;
   fi
 - if [ "$TRAVIS_OS_NAME" == "windows" ]; then curl -0 https://raw.githubusercontent.com/NREL/EnergyPlus/3cf5e1c8e6944e8a7760b80078c6945073cc8364/cmake/qtifw/install_script.qs -o install_script.qs;
-  $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs; dir;
+  start /wait "" $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs; dir;
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $DEST;
   powershell -command Expand-Archive -Path $ATTCHNUM.zip -DestinationPath $DEST -Force; fi
-- curl -SL https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.sh -o miniconda.sh
-- bash miniconda.sh -b -p $HOME/miniconda
+- curl -SLO https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT
+- if [ "$TRAVIS_OS_NAME" == "linux" ]; then bash miniconda.sh -b -p $HOME/miniconda; fi
+- if [ "$TRAVIS_OS_NAME" == "osx" ]; then bash miniconda.sh -b -p $HOME/miniconda; fi
+- if [ "$TRAVIS_OS_NAME" == "windows" ]; then start /wait "" Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT
 - source $HOME/miniconda/etc/profile.d/conda.sh
 - conda config --set always_yes yes --set changeps1 no
 - conda update -q conda;
