+++ .travis.yml
@@ -81,7 +81,7 @@ before_install:
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $DEST;
   powershell -command Expand-Archive -Path $ATTCHNUM.zip -DestinationPath $DEST -Force; fi
-- curl -SLO https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT
+- curl -SL https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT -o Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT
 - if [ "$TRAVIS_OS_NAME" == "linux" ]; then bash miniconda.sh -b -p $HOME/miniconda; fi
 - if [ "$TRAVIS_OS_NAME" == "osx" ]; then bash miniconda.sh -b -p $HOME/miniconda; fi
 - if [ "$TRAVIS_OS_NAME" == "windows" ]; then start /wait "" Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT
