+++ .travis.yml
@@ -12,16 +12,16 @@ jobs:
   include:
     - name: Python 3.8 on Xenial Linux
       python: 3.8           # this works for Linux but is ignored on macOS or Windows
-      env: MINICONDA_OS=Linux; MINICONDAEXT=sh
+      env: MINICONDA_OS=Linux
     - name: Python 3.8 on macOS
       os: osx
       osx_image: xcode11.2
       language: objective-c
-      env: PYENV_VERSION=3.8; MINICONDA_OS=MacOSX; MINICONDAEXT=sh
+      env: PYENV_VERSION=3.8; MINICONDA_OS=MacOSX
     - name: "Python 3.8 on Windows"
       os: windows           # Windows 10.0.17134 N/A Build 17134
       language: shell       # 'language: python' is an error on Travis CI Windows
-      env: PATH=/c/Python38:/c/Python38/Scripts:$PATH; MINICONDA_OS=Windows; MINICONDAEXT=exe
+      env: PATH=/c/Python38:/c/Python38/Scripts:$PATH; MINICONDA_OS=Windows
     - stage: deploy
       script: echo "Deploying to PyPi and GitHub releases ..."
       deploy:
@@ -81,11 +81,12 @@ before_install:
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $DEST;
   powershell -command Expand-Archive -Path $ATTCHNUM.zip -DestinationPath $DEST -Force; fi
-- curl -SL https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT -o Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT
-- if [ "$TRAVIS_OS_NAME" == "linux" ]; then bash miniconda.sh -b -p $HOME/miniconda; fi
-- if [ "$TRAVIS_OS_NAME" == "osx" ]; then bash miniconda.sh -b -p $HOME/miniconda; fi
-- if [ "$TRAVIS_OS_NAME" == "windows" ]; then start Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.$MINICONDAEXT; fi
-- source $HOME/miniconda/etc/profile.d/conda.sh
+- curl -SL https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.sh -o miniconda.sh
+- if [ "$TRAVIS_OS_NAME" == "linux" ]; then bash miniconda.sh -b -p "$HOME"/miniconda; source
+  $HOME/miniconda/etc/profile.d/conda.sh; fi
+- if [ "$TRAVIS_OS_NAME" == "osx" ]; then bash miniconda.sh -b -p "$HOME"/miniconda; source
+  ME/miniconda/etc/profile.d/conda.sh; fi
+- if [ "$TRAVIS_OS_NAME" == "windows" ]; then choco install miniconda3; fi
 - conda config --set always_yes yes --set changeps1 no
 - conda update -q conda;
 - conda info -a
