+++ .travis.yml
@@ -86,7 +86,8 @@ before_install:
   $HOME/miniconda/etc/profile.d/conda.sh; fi
 - if [ "$TRAVIS_OS_NAME" == "osx" ]; then bash miniconda.sh -b -p "$HOME"/miniconda; source
   $HOME/miniconda/etc/profile.d/conda.sh; fi
-- if [ "$TRAVIS_OS_NAME" == "windows" ]; then choco install miniconda3 --params="'/AddToPath:1'"; fi
+- if [ "$TRAVIS_OS_NAME" == "windows" ]; then wget https://raw.githubusercontent
+  .com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh; source conda4travis.sh; fi
 - conda config --set always_yes yes --set changeps1 no
 - conda update -q conda;
 - conda info -a
