### 2020-04-15 493da85
Simplifies miniconda install
