+++ .travis.yml
@@ -81,12 +81,7 @@ before_install:
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $DEST;
   powershell -command Expand-Archive -Path $ATTCHNUM.zip -DestinationPath $DEST -Force; fi
-- curl -SL https://repo.anaconda.com/miniconda/Miniconda3-$MINICONDA_VERSION-$MINICONDA_OS-x86_64.sh -o miniconda.sh
-- if [ "$TRAVIS_OS_NAME" == "linux" ]; then bash miniconda.sh -b -p "$HOME"/miniconda; source
-  $HOME/miniconda/etc/profile.d/conda.sh; fi
-- if [ "$TRAVIS_OS_NAME" == "osx" ]; then bash miniconda.sh -b -p "$HOME"/miniconda; source
-  $HOME/miniconda/etc/profile.d/conda.sh; fi
-- if [ "$TRAVIS_OS_NAME" == "windows" ]; then wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh; source conda4travis.sh; fi
+- wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh; source conda4travis.sh; fi
 - conda config --set always_yes yes --set changeps1 no
 - conda update -q conda;
 - conda info -a
