+++ .travis.yml
@@ -81,9 +81,10 @@ before_install:
   DEST=C:\\EnergyPlusV$ENERGYPLUS_INSTALL_VERSION\\PreProcess\IDFVersionUpdater; echo "Extracting and Copying files to
   ..."; echo $DEST;
   powershell -command Expand-Archive -Path $ATTCHNUM.zip -DestinationPath $DEST -Force; fi
-- wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh; source conda4travis.sh; fi
+- wget https://raw.githubusercontent.com/trichter/conda4travis/latest/conda4travis.sh -O conda4travis.sh
+- source conda4travis.sh
 - conda config --set always_yes yes --set changeps1 no
-- conda update -q conda;
+- conda update -q conda
 - conda info -a
 - conda env update -q -n test-env python=$TRAVIS_PYTHON_VERSION -f environment.yml
 - conda activate test-env
