+++ .travis.yml
@@ -90,9 +90,9 @@ before_install:
 - conda activate test-env
 - python --version
 install:
-- pip3 install --upgrade pip
-- pip3 install --upgrade setuptools
-- pip3 install .[dev]
+- pip install --upgrade pip
+- pip install --upgrade setuptools
+- pip install .[dev]
 after_success:
 - coverage report -m
 - coveralls
