+++ .travis.yml
@@ -12,16 +12,13 @@ jobs:
   include:
     - name: Python 3.8 on Xenial Linux
       python: 3.8           # this works for Linux but is ignored on macOS or Windows
-      env: MINICONDA_OS=Linux
     - name: Python 3.8 on macOS
       os: osx
       osx_image: xcode11.2
       language: objective-c
-      env: PYENV_VERSION=3.8; MINICONDA_OS=MacOSX
     - name: "Python 3.8 on Windows"
       os: windows           # Windows 10.0.17134 N/A Build 17134
       language: shell       # 'language: python' is an error on Travis CI Windows
-      env: PATH=/c/Python38:/c/Python38/Scripts:$PATH; MINICONDA_OS=Windows
     - stage: deploy
       script: echo "Deploying to PyPi and GitHub releases ..."
       deploy:
