+++ .travis.yml
@@ -12,6 +12,7 @@ jobs:
   include:
     - name: Python 3.8 on Xenial Linux
       python: 3.8           # this works for Linux but is ignored on macOS or Windows
+      script: py.test --cov=archetypal --verbose tests/
       after_success:
         - coverage report -m
         - coveralls
@@ -19,10 +20,11 @@ jobs:
       os: osx
       osx_image: xcode11.2
       language: objective-c
+      script: py.test --cov=archetypal --verbose tests/
     - name: "Python 3.8 on Windows"
       os: windows           # Windows 10.0.17134 N/A Build 17134
       language: shell       # 'language: python' is an error on Travis CI Windows
-      env: YARN_GPG=no      # https://travis-ci.community/t/timeout-after-build-finished-and-succeeded/1336
+      script: py.test --cov=archetypal --verbose tests/
     - stage: deploy
       script: echo "Deploying to PyPi and GitHub releases ..."
       deploy:
@@ -92,4 +94,3 @@ install:
 - pip install --upgrade pip
 - pip install --upgrade setuptools
 - pip install .[dev]
-script: echo "Running Tests ..." # stage: test
