### 2020-04-16 fb83613
Replaces energyplus install with script
