### 2020-04-16 cad7192
No need to remove install script
