### 2020-04-16 97323c1
Checks if IDD file exists
