### 2020-04-16 b34f157
getting latest install script
