+++ .travis.yml
@@ -1,8 +1,8 @@
 language: python
 env:
   global:
-  - ARCHETYPAL_INTEGRATION=True ENERGYPLUS_VERSION=9.2.0 ENERGYPLUS_SHA=921312fa1d
-    ENERGYPLUS_INSTALL_VERSION=9-2-0 MINICONDA_VERSION=latest
+  - ARCHETYPAL_INTEGRATION=True ENERGYPLUS_VERSION=9.3.0 ENERGYPLUS_SHA=baff08990c
+    ENERGYPLUS_INSTALL_VERSION=9-3-0 MINICONDA_VERSION=latest
 addons:
   apt:
     packages:
