### 2020-04-16 896b1f5
Revert "Try EnergyPlus 9.3.0"
This reverts commit 6e93befebda6dc516dc97303559fec51b5846ce9.
