### 2020-04-16 8bf6f26
switching to zip install
