+++ install_energyplus.sh
@@ -17,7 +17,7 @@ if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
   PLATFORM=Linux
 fi
 if [[ "$TRAVIS_OS_NAME" == "windows" ]]; then
-  EXT=exe
+  EXT=zip
   PLATFORM=Windows
 fi
 # Download EnergyPlus executable
@@ -68,16 +68,11 @@ if [ "$TRAVIS_OS_NAME" == "osx" ]; then
   sudo rm $ATTCHNUM.zip
 fi
 if [ "$TRAVIS_OS_NAME" == "windows" ]; then
-  # getting custom install script https://github.com/NREL/EnergyPlus/pull/7615
-  curl -SL -C - https://raw.githubusercontent.com/jmarrec/EnergyPlus/40afb275f66201db5305f54df6c070d0b0cb4fc3/cmake/qtifw/install_script.qs -o install_script.qs
-  ./$ENERGYPLUS_DOWNLOAD_FILENAME.$EXT --verbose --script install_script.qs
-  DEST=C:\\EnergyPlusV"$ENERGYPLUS_INSTALL_VERSION"\\PreProcess\\IDFVersionUpdater
+  DEST=C:\\EnergyPlusV"$ENERGYPLUS_INSTALL_VERSION"
   echo "Extracting and Copying files to... $DEST"
-  powershell Expand-Archive -Path $ATTCHNUM.zip -DestinationPath "$DEST" -Force
+  powershell Expand-Archive -Path $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT -DestinationPath "$DEST" -Force
   # cleanup
-  rm -v install_script.qs
   rm -v $ENERGYPLUS_DOWNLOAD_FILENAME.$EXT
-  rm -v $ATTCHNUM.zip
   IDD=C:\\EnergyPlusV"$ENERGYPLUS_INSTALL_VERSION"\\Energy+.idd
   if [ -f "$IDD" ]; then
     echo "$IDD" exists
