### 2020-04-16 9dc1c4f
Pypi on master only + github releases on all branches
