+++ .travis.yml
@@ -40,7 +40,7 @@ jobs:
         cleanup: false
       - provider: releases
         api_key:
-          secure: fnWLQlgUCcRD+q6Js/IgRnAQYPqoLrQfGlvqBpwhH/iO88UKmXp1hco4dnHSditFUy23QmLwSNzS4cb1GzoDnKnH022+u5GgL2NDxqm3lB5jEuKvUlMA59mJiSbr84Of2RAlHAwPTj0Ux7esq8tPyjipKJkdiKAX85/lCuEOOBs18WbJaBzbYrYiw6dTGT57xCOB7e+0CDAwdTVBB7388rvbCgGkTji7hC/jkThjkVqnPFXNcH/6ehcqE6vCkJzlT9tf1AypweWjifoyW3e5H7vKWrEt0v+GJfEG81SekaNhXpFcm6mACBfcGUNMDoNwsCpp/de93o3oQl/OYyuZ7tB2FcDzIJ4nytJNEGxs6ZN8tDgcMaKU/H6FEpVskrwQElGPK72SXnhmv9lX4e2Rm2ku4jt7rFukn+cFrkWkKDzP64o2giAn7a1yaWpMvAVPhmaT8eX6am8lqzOIESQDpspFaNp0owYO/mj3LVItKC0fsuCANuyUFZx98AoNuyXY8hXtF4sd/Oq+UcgLAVXU5KljSQ9KZi7XNvTh953OtB2O/jJyC+IbQGQaNJB+TTimvHBFohUTM18uXMP+2bcAGzv4b6uyh3N5u2oRVaeQ115fTmnJibQVkvMsPKm95QiysTNqhA/rNMRo8dz/DW6SdjzlqTE4LWJjZqjbsmKh0hQ=
+          secure: hAbE9izaOts/fDGHD/l+8I6rcB9ax8QEepdL7VNCtU9Ta3geY64P+G3mh3EDK3dKMrtoLbpcgkP5XobUz3MP7ifPBxbXYE+AHF74oMyxGnqAjAl51drMbrgd+veCOd1k/0xdqK/GOxAg6yrPgUsW2hqiEueB/CG9NV+U/cRXo5HBhNPjuA52Ea+9FCEM0Ufd6m+QyQw8HOrJlpNQYWxctPw2hULmmR+wphdEwQxSU+E5Cfj3yEAfOBS6vMwysItSG9T0Is1E4yw8NC9tJDodpVMyxgp4V/4iMqX5txnycaynP1yxPBEVciY4pwAHWsSy51UVL83lUFqc/zOYCV7XD3mNHA7pTs3uRAsI+Ac1J9iqDqLt8GitE89vKq2hExMFPi+F7EXpvrRHFF5asfOb/RJpt5LH0fAVgPLNcWZS1vsXSpQ2N6Ws//2V4hzx8kiYP/d3GQtW7aSU9EzRp/Sl/WIQz0dNeTsTYQiHrgOMrWo8SV9RO2enghlSUA5aep9gqA7zhjXYQJNtqPUdWSi3UCENKD87P4FrerHR+uP2C7tpuWbi/76RFA7XWm9nZPkhQR7eDOGnWw2doHt6DYM3cMUI68udThiuTb8IgNig9JLrOiZzI7n8ndffUJrumznMh9Ud4mn38eFVm+YA+qWjJyMljqKCOBSt7b5xBq9eE3Q=
         file_glob: true
         file:
           - dist/*.whl
