### 2020-04-16 ebe6058
directory will not be cleaned up before the deploy step
