### 2020-04-16 9c9cc32
Revert "Comments the script"
This reverts commit 98752fff8a6ca35519f9be7bb832f8290ac440aa.
