### 2020-04-16 e5d8efa
Puts back extra downloads on windows build
