### 2020-04-16 d6a5b24
Implements travis-ci build for Linus, macOS and Windows (#76)
* Adds deployment build to pypi

* Sets up stages

* Test: comments test script

* Organizes builds + adds Github Releases deploy

* tests

* Trying to fix test stage

* Another way

* Fixes pip error

* Try fix

* Should fix windows build

* Install package using --user

* Removes "" in HOME

* Moves python before_install in global before_install

* Removes --user

* Fixes version

* Setting energyplus install for travis windows

* Moved remove

* Reformat

* Fix format

* typo

* Missing demi-colon

* env path

* Powershell

* Try
[…]
