+++ .github/release-drafter.yml
@@ -8,7 +8,7 @@ template: |
   $CONTRIBUTORS
 
 categories:
-  - title: 'Features'
+  - title: 'New Features'
     labels:
       - 'feature'
       - 'enhancement'
@@ -17,5 +17,8 @@ categories:
       - 'fix'
       - 'bugfix'
       - 'bug'
+  - title: 'Under the hood refactoring ⚙'
+    labels:
+      - 'refactoring'
 name-template: 'v$NEXT_PATCH_VERSION 🌈'
 tag-template: 'v$NEXT_PATCH_VERSION'
