### 2020-04-16 a1141cd
Refactors tests to be faster
