+++ tests/test_trnsys.py
@@ -61,6 +61,356 @@ from tests.conftest import get_platform
 
 
 @pytest.fixture(
+    scope="class", params=["tests/input_data/trnsys/simple_2_zone_sched.idf",],
+)
+def converttesteasy(request):
+    file = request.param
+    window_file = "W74-lib.dat"
+    template_dir = os.path.join("archetypal", "ressources")
+    window_filepath = os.path.join(template_dir, window_file)
+    template_d18 = "tests/input_data/trnsys/NewFileTemplate.d18"
+    trnsidf_exe = "docker/trnsidf/trnsidf.exe"  # 'docker/trnsidf/trnsidf.exe'
+
+    # prepare args (key=value). Key is a unique id for the runs (here the
+    # file basename is used). Value is a dict of the function arguments
+    kwargs_dict = {
+        "u_value": 2.5,
+        "shgc": 0.6,
+        "t_vis": 0.78,
+        "tolerance": 0.05,
+        "ordered": True,
+    }
+    idf = load_idf(file)
+
+    weather_file = os.path.join(
+        "tests", "input_data", "CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw"
+    )
+
+    yield idf, file, weather_file, window_filepath, trnsidf_exe, template_d18, kwargs_dict
+
+    del idf
+
+
+class TestConvertEasy:
+
+    """Tests convert_idf_to_trnbuild() 1 file"""
+
+    def test_get_save_write_schedules_as_sched(self, config, converttesteasy):
+        output_folder = None
+        (
+            idf,
+            idf_file,
+            weather_file,
+            window_lib,
+            trnsidf_exe,
+            template,
+            _,
+        ) = converttesteasy
+        lines = io.TextIOWrapper(io.BytesIO(settings.template_BUI)).readlines()
+        try:
+            (
+                idf_file,
+                weather_file,
+                window_lib,
+                output_folder,
+                trnsidf_exe,
+                template,
+            ) = _assert_files(
+                idf_file, weather_file, window_lib, output_folder, trnsidf_exe, template
+            )
+        except:
+            output_folder = os.path.relpath(settings.data_folder)
+            print("Could not assert all paths exist - OK for this test")
+        schedule_names, schedules = _get_schedules(idf)
+        _yearlySched_to_csv(idf_file, output_folder, schedule_names, schedules)
+        schedule_as_input = False
+        schedules_not_written = _write_schedules(
+            lines, schedule_names, schedules, schedule_as_input, idf_file
+        )
+
+    def test_write_version_and_building(self, config, converttesteasy):
+        (
+            idf,
+            idf_file,
+            weather_file,
+            window_lib,
+            trnsidf_exe,
+            template,
+            _,
+        ) = converttesteasy
+        (
+            buildingSurfs,
+            buildings,
+            constructions,
+            equipments,
+            fenestrationSurfs,
+            globGeomRules,
+            lights,
+            locations,
+            materialAirGap,
+            materialNoMass,
+            materials,
+            peoples,
+            versions,
+            zones,
+            zonelists,
+        ) = get_idf_objects(idf)
+        lines = io.TextIOWrapper(io.BytesIO(settings.template_BUI)).readlines()
+        _write_version(lines, versions)
+        _write_building(buildings, lines)
+
+    def test_write_material(self, config, converttesteasy):
+        (
+            idf,
+            idf_file,
+            weather_file,
+            window_lib,
+            trnsidf_exe,
+            template,
+            _,
+        ) = converttesteasy
+
+        # Read IDF_T3D template and write lines in variable
+        lines = io.TextIOWrapper(io.BytesIO(settings.template_BUI)).readlines()
+
+        # Get objects from IDF file
[…]
