### 2020-04-17 0f8a6bd
Better coverage for test_write_to_b18()
