### 2020-04-17 68a45cc
Asserts test_load_idf_file_and_clean_names()
Makes sure load_idf_file_and_clean_names() returns unique names and an IDF object
