### 2020-04-17 f81e429
Asserts test_get_save_write_schedules_as_sched()
Makes sure csv file with schedules exists and that schedules are written in lines
