### 2020-04-17 fdcb2be
Asserts test_write_version_and_building()
Makes sure version and building information written in lines
