### 2020-04-17 8f9f41b
Enhances test_get_save_write_schedules_as_sched() assertion
