### 2020-04-17 db07fe4
Asserts test_write_material()
Makes sure materials (material, AirGap, NoMass, etc.) are written in lines
