### 2020-04-17 8cedf76
Asserts test_relative_to_absolute()
Makes sure relative coords converted to absolute ones
