### 2020-04-17 3fd37e2
Asserts test_save_t3d()
Makes sure path to T3D file exists
