### 2020-04-20 #1833 @kyleniemeyer
Hi @samuelduchesne, looks like we just need a few edits to the paper. Please let me know when you've done that, and I'll do my final checks.
