### 2020-05-20 #1833 @arfon
:wave: @samuelduchesne - JOSS [is now open again](https://blog.joss.theoj.org/2020/05/reopening-joss) for submissions and we're actively trying to wrap up existing reviews.

Do you think you might be able to complete these last few items soon?
