### 2020-05-28 #1833 @samuelduchesne
@brynpickering, I have addressed the caching in https://github.com/samuelduchesne/archetypal/pull/87
