### 2020-06-07 #1833 @kyleniemeyer
Hi @samuelduchesne, I think we're ready to wrap things up here.

Can you make these final small edits to the paper?
- In the summary paragraph, the sentence
> archetypal is a Python package that helps handling collections of such archetypes and to enabling the interoperability between these energy simulation platforms to accelerate the creation of reliable urban building energy models.
seems incorrect, because of the "and to enabling" (it's also a bit long and complicated). I think changing that to "to enable" helps fix it, but you might consider revising the sentence overall for clarify.
- in the paragraph about caching, I think API should be capitalized.

Once you've made those changes, please archive your repository on Zenodo and provide the DOI here.
