### 2020-06-08 #1833 @samuelduchesne
Hi @kyleniemeyer, thanks for the feedback. Paper updated with https://github.com/samuelduchesne/archetypal/commit/e9232f26c83247249b7d12d228e53480bcf4cf19 and following DOI for zenodo files: [10.5281/zenodo.3885587](10.5281/zenodo.3885587)
