### 2020-06-08 #1833 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
